@@ -18,6 +18,7 @@ import (
 	"github.com/Mrs4s/MiraiGo/client"
 	"github.com/Mrs4s/MiraiGo/message"
 	"github.com/Mrs4s/MiraiGo/utils"
+	"github.com/segmentio/asm/base64"
 	log "github.com/sirupsen/logrus"
 	"github.com/tidwall/gjson"
 
@@ -25,14 +26,31 @@ import (
 	"github.com/Mrs4s/go-cqhttp/global"
 	"github.com/Mrs4s/go-cqhttp/internal/base"
 	"github.com/Mrs4s/go-cqhttp/internal/cache"
+	"github.com/Mrs4s/go-cqhttp/internal/download"
+	"github.com/Mrs4s/go-cqhttp/internal/msg"
 	"github.com/Mrs4s/go-cqhttp/internal/param"
 	"github.com/Mrs4s/go-cqhttp/modules/filter"
+	"github.com/Mrs4s/go-cqhttp/pkg/onebot"
 )
+
+type guildMemberPageToken struct {
+	guildID        uint64
+	nextIndex      uint32
+	nextRoleID     uint64
+	nextQueryParam string
+}
+
+var defaultPageToken = guildMemberPageToken{
+	guildID:    0,
+	nextIndex:  0,
+	nextRoleID: 2,
+}
 
 // CQGetLoginInfo 获取登录号信息
 //
 // https://git.io/Jtz1I
-// @route(get_login_info)
+// @route11(get_login_info)
+// @route12(get_self_info)
 func (bot *CQBot) CQGetLoginInfo() global.MSG {
 	return OK(global.MSG{"user_id": bot.Client.Uin, "nickname": bot.Client.Nickname})
 }
@@ -116,7 +134,7 @@ func (bot *CQBot) CQGetGuildChannelList(guildID uint64, noCache bool) global.MSG
 	if noCache {
 		channels, err := bot.Client.GuildService.FetchChannelList(guildID)
 		if err != nil {
-			log.Errorf("获取频道 %v 子频道列表时出现错误: %v", guildID, err)
+			log.Warnf("获取频道 %v 子频道列表时出现错误: %v", guildID, err)
 			return Failed(100, "API_ERROR", err.Error())
 		}
 		guild.Channels = channels
@@ -128,24 +146,83 @@ func (bot *CQBot) CQGetGuildChannelList(guildID uint64, noCache bool) global.MSG
 	return OK(channels)
 }
 
-/*
 // CQGetGuildMembers 获取频道成员列表
-// @route(get_guild_members)
-func (bot *CQBot) CQGetGuildMembers(guildID uint64) global.MSG {
+// @route(get_guild_member_list)
+func (bot *CQBot) CQGetGuildMembers(guildID uint64, nextToken string) global.MSG {
 	guild := bot.Client.GuildService.FindGuild(guildID)
 	if guild == nil {
 		return Failed(100, "GUILD_NOT_FOUND")
 	}
-	return OK(nil) // todo
+	token := &defaultPageToken
+	if nextToken != "" {
+		i, exists := bot.nextTokenCache.Get(nextToken)
+		if !exists {
+			return Failed(100, "NEXT_TOKEN_NOT_EXISTS")
+		}
+		token = i
+		if token.guildID != guildID {
+			return Failed(100, "GUILD_NOT_MATCH")
+		}
+	}
+	ret, err := bot.Client.GuildService.FetchGuildMemberListWithRole(guildID, 0, token.nextIndex, token.nextRoleID, token.nextQueryParam)
+	if err != nil {
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	res := global.MSG{
+		"members":    convertGuildMemberInfo(ret.Members),
+		"finished":   ret.Finished,
+		"next_token": nil,
+	}
+	if !ret.Finished {
+		next := &guildMemberPageToken{
+			guildID:        guildID,
+			nextIndex:      ret.NextIndex,
+			nextRoleID:     ret.NextRoleId,
+			nextQueryParam: ret.NextQueryParam,
+		}
+		id := base64.StdEncoding.EncodeToString(binary.NewWriterF(func(w *binary.Writer) {
+			w.WriteUInt64(uint64(time.Now().UnixNano()))
+			w.WriteString(utils.RandomString(5))
+		}))
+		bot.nextTokenCache.Add(id, next, time.Minute*10)
+		res["next_token"] = id
+	}
+	return OK(res)
 }
-*/
+
+// CQGetGuildMemberProfile 获取频道成员资料
+// @route(get_guild_member_profile)
+func (bot *CQBot) CQGetGuildMemberProfile(guildID, userID uint64) global.MSG {
+	if bot.Client.GuildService.FindGuild(guildID) == nil {
+		return Failed(100, "GUILD_NOT_FOUND")
+	}
+	profile, err := bot.Client.GuildService.FetchGuildMemberProfileInfo(guildID, userID)
+	if err != nil {
+		log.Warnf("获取频道 %v 成员 %v 资料时出现错误: %v", guildID, userID, err)
+		return Failed(100, "API_ERROR", err.Error())
+	}
+	roles := make([]global.MSG, 0, len(profile.Roles))
+	for _, role := range profile.Roles {
+		roles = append(roles, global.MSG{
+			"role_id":   fU64(role.RoleId),
+			"role_name": role.RoleName,
+		})
+	}
+	return OK(global.MSG{
+		"tiny_id":    fU64(profile.TinyId),
+		"nickname":   profile.Nickname,
+		"avatar_url": profile.AvatarUrl,
+		"join_time":  profile.JoinTime,
+		"roles":      roles,
+	})
+}
 
 // CQGetGuildRoles 获取频道角色列表
 // @route(get_guild_roles)
 func (bot *CQBot) CQGetGuildRoles(guildID uint64) global.MSG {
 	r, err := bot.Client.GuildService.GetGuildRoles(guildID)
 	if err != nil {
-		log.Errorf("获取频道 %v 角色列表时出现错误: %v", guildID, err)
+		log.Warnf("获取频道 %v 角色列表时出现错误: %v", guildID, err)
 		return Failed(100, "API_ERROR", err.Error())
 	}
 	roles := make([]global.MSG, len(r))
@@ -175,7 +252,7 @@ func (bot *CQBot) CQCreateGuildRole(guildID uint64, name string, color uint32, i
 	}
 	role, err := bot.Client.GuildService.CreateGuildRole(guildID, name, color, independent, userSlice)
 	if err != nil {
-		log.Errorf("创建频道 %v 角色时出现错误: %v", guildID, err)
+		log.Warnf("创建频道 %v 角色时出现错误: %v", guildID, err)
 		return Failed(100, "API_ERROR", err.Error())
 	}
 	return OK(global.MSG{
@@ -188,7 +265,7 @@ func (bot *CQBot) CQCreateGuildRole(guildID uint64, name string, color uint32, i
 func (bot *CQBot) CQDeleteGuildRole(guildID uint64, roleID uint64) global.MSG {
 	err := bot.Client.GuildService.DeleteGuildRole(guildID, roleID)
 	if err != nil {
-		log.Errorf("删除频道 %v 角色时出现错误: %v", guildID, err)
+		log.Warnf("删除频道 %v 角色时出现错误: %v", guildID, err)
 		return Failed(100, "API_ERROR", err.Error())
 	}
 	return OK(nil)
@@ -205,7 +282,7 @@ func (bot *CQBot) CQSetGuildMemberRole(guildID uint64, set bool, roleID uint64, 
 	}
 	err := bot.Client.GuildService.SetUserRoleInGuild(guildID, set, roleID, userSlice)
 	if err != nil {
-		log.Errorf("设置用户在频道 %v 中的角色时出现错误: %v", guildID, err)
+		log.Warnf("设置用户在频道 %v 中的角色时出现错误: %v", guildID, err)
 		return Failed(100, "API_ERROR", err.Error())
 	}
 	return OK(nil)
@@ -216,7 +293,7 @@ func (bot *CQBot) CQSetGuildMemberRole(guildID uint64, set bool, roleID uint64, 
 func (bot *CQBot) CQModifyRoleInGuild(guildID uint64, roleID uint64, name string, color uint32, indepedent bool) global.MSG {
 	err := bot.Client.GuildService.ModifyRoleInGuild(guildID, roleID, name, color, indepedent)
 	if err != nil {
-		log.Errorf("修改频道 %v 角色时出现错误: %v", guildID, err)
+		log.Warnf("修改频道 %v 角色时出现错误: %v", guildID, err)
 		return Failed(100, "API_ERROR", err.Error())
 	}
 	return OK(nil)
@@ -238,7 +315,7 @@ func (bot *CQBot) CQGetTopicChannelFeeds(guildID, channelID uint64) global.MSG {
 	}
 	feeds, err := bot.Client.GuildService.GetTopicChannelFeeds(guildID, channelID)
 	if err != nil {
-		log.Errorf("获取频道 %v 帖子时出现错误: %v", channelID, err)
+		log.Warnf("获取频道 %v 帖子时出现错误: %v", channelID, err)
 		return Failed(100, "API_ERROR", err.Error())
 	}
 	c := make([]global.MSG, 0, len(feeds))
@@ -252,13 +329,13 @@ func (bot *CQBot) CQGetTopicChannelFeeds(guildID, channelID uint64) global.MSG {
 //
 // https://git.io/Jtz1L
 // @route(get_friend_list)
-func (bot *CQBot) CQGetFriendList() global.MSG {
+func (bot *CQBot) CQGetFriendList(spec *onebot.Spec) global.MSG {
 	fs := make([]global.MSG, 0, len(bot.Client.FriendList))
 	for _, f := range bot.Client.FriendList {
 		fs = append(fs, global.MSG{
 			"nickname": f.Nickname,
 			"remark":   f.Remark,
-			"user_id":  f.Uin,
+			"user_id":  spec.ConvertID(f.Uin),
 		})
 	}
 	return OK(fs)
@@ -270,7 +347,7 @@ func (bot *CQBot) CQGetFriendList() global.MSG {
 func (bot *CQBot) CQGetUnidirectionalFriendList() global.MSG {
 	list, err := bot.Client.GetUnidirectionalFriendList()
 	if err != nil {
-		log.Errorf("获取单向好友列表时出现错误: %v", err)
+		log.Warnf("获取单向好友列表时出现错误: %v", err)
 		return Failed(100, "API_ERROR", err.Error())
 	}
 	fs := make([]global.MSG, 0, len(list))
@@ -291,13 +368,13 @@ func (bot *CQBot) CQGetUnidirectionalFriendList() global.MSG {
 func (bot *CQBot) CQDeleteUnidirectionalFriend(uin int64) global.MSG {
 	list, err := bot.Client.GetUnidirectionalFriendList()
 	if err != nil {
-		log.Errorf("获取单向好友列表时出现错误: %v", err)
+		log.Warnf("获取单向好友列表时出现错误: %v", err)
 		return Failed(100, "API_ERROR", err.Error())
 	}
 	for _, f := range list {
 		if f.Uin == uin {
 			if err = bot.Client.DeleteUnidirectionalFriend(uin); err != nil {
-				log.Errorf("删除单向好友时出现错误: %v", err)
+				log.Warnf("删除单向好友时出现错误: %v", err)
 				return Failed(100, "API_ERROR", err.Error())
 			}
 			return OK(nil)
@@ -314,7 +391,7 @@ func (bot *CQBot) CQDeleteFriend(uin int64) global.MSG {
 		return Failed(100, "FRIEND_NOT_FOUND", "好友不存在")
 	}
 	if err := bot.Client.DeleteFriend(uin); err != nil {
-		log.Errorf("删除好友时出现错误: %v", err)
+		log.Warnf("删除好友时出现错误: %v", err)
 		return Failed(100, "DELETE_API_ERROR", err.Error())
 	}
 	return OK(nil)
@@ -324,16 +401,15 @@ func (bot *CQBot) CQDeleteFriend(uin int64) global.MSG {
 //
 // https://git.io/Jtz1t
 // @route(get_group_list)
-func (bot *CQBot) CQGetGroupList(noCache bool) global.MSG {
+func (bot *CQBot) CQGetGroupList(noCache bool, spec *onebot.Spec) global.MSG {
 	gs := make([]global.MSG, 0, len(bot.Client.GroupList))
 	if noCache {
 		_ = bot.Client.ReloadGroupList()
 	}
 	for _, g := range bot.Client.GroupList {
 		gs = append(gs, global.MSG{
-			"group_id":          g.Code,
+			"group_id":          spec.ConvertID(g.Code),
 			"group_name":        g.Name,
-			"group_memo":        g.Memo,
 			"group_create_time": g.GroupCreateTime,
 			"group_level":       g.GroupLevel,
 			"max_member_count":  g.MaxMemberCount,
@@ -347,7 +423,7 @@ func (bot *CQBot) CQGetGroupList(noCache bool) global.MSG {
 //
 // https://git.io/Jtz1O
 // @route(get_group_info)
-func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) global.MSG {
+func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool, spec *onebot.Spec) global.MSG {
 	group := bot.Client.FindGroup(groupID)
 	if group == nil || noCache {
 		group, _ = bot.Client.GetGroupInfo(groupID)
@@ -361,7 +437,7 @@ func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) global.MSG {
 		for _, g := range info {
 			if g.Code == groupID {
 				return OK(global.MSG{
-					"group_id":          g.Code,
+					"group_id":          spec.ConvertID(g.Code),
 					"group_name":        g.Name,
 					"group_memo":        g.Memo,
 					"group_create_time": 0,
@@ -373,9 +449,8 @@ func (bot *CQBot) CQGetGroupInfo(groupID int64, noCache bool) global.MSG {
 		}
 	} else {
 		return OK(global.MSG{
-			"group_id":          group.Code,
+			"group_id":          spec.ConvertID(group.Code),
 			"group_name":        group.Name,
-			"group_memo":        group.Memo,
 			"group_create_time": group.GroupCreateTime,
 			"group_level":       group.GroupLevel,
 			"max_member_count":  group.MaxMemberCount,
@@ -442,7 +517,7 @@ func (bot *CQBot) CQGetGroupMemberInfo(groupID, userID int64, noCache bool) glob
 func (bot *CQBot) CQGetGroupFileSystemInfo(groupID int64) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	return OK(fs)
@@ -455,12 +530,12 @@ func (bot *CQBot) CQGetGroupFileSystemInfo(groupID int64) global.MSG {
 func (bot *CQBot) CQGetGroupRootFiles(groupID int64) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	files, folders, err := fs.Root()
 	if err != nil {
-		log.Errorf("获取群 %v 根目录文件失败: %v", groupID, err)
+		log.Warnf("获取群 %v 根目录文件失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	return OK(global.MSG{
@@ -476,12 +551,12 @@ func (bot *CQBot) CQGetGroupRootFiles(groupID int64) global.MSG {
 func (bot *CQBot) CQGetGroupFilesByFolderID(groupID int64, folderID string) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	files, folders, err := fs.GetFilesByFolder(folderID)
 	if err != nil {
-		log.Errorf("获取群 %v 根目录 %v 子文件失败: %v", groupID, folderID, err)
+		log.Warnf("获取群 %v 根目录 %v 子文件失败: %v", groupID, folderID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	return OK(global.MSG{
@@ -494,6 +569,7 @@ func (bot *CQBot) CQGetGroupFilesByFolderID(groupID int64, folderID string) glob
 //
 // https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96%E7%BE%A4%E6%96%87%E4%BB%B6%E8%B5%84%E6%BA%90%E9%93%BE%E6%8E%A5
 // @route(get_group_file_url)
+// @rename(bus_id->"[busid\x2Cbus_id].0")
 func (bot *CQBot) CQGetGroupFileURL(groupID int64, fileID string, busID int32) global.MSG {
 	url := bot.Client.GetGroupFileUrl(groupID, fileID, busID)
 	if url == "" {
@@ -510,19 +586,44 @@ func (bot *CQBot) CQGetGroupFileURL(groupID int64, fileID string, busID int32) g
 // @route(upload_group_file)
 func (bot *CQBot) CQUploadGroupFile(groupID int64, file, name, folder string) global.MSG {
 	if !global.PathExists(file) {
-		log.Errorf("上传群文件 %v 失败: 文件不存在", file)
+		log.Warnf("上传群文件 %v 失败: 文件不存在", file)
 		return Failed(100, "FILE_NOT_FOUND", "文件不存在")
 	}
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	if folder == "" {
 		folder = "/"
 	}
 	if err = fs.UploadFile(file, name, folder); err != nil {
-		log.Errorf("上传群 %v 文件 %v 失败: %v", groupID, file, err)
+		log.Warnf("上传群 %v 文件 %v 失败: %v", groupID, file, err)
+		return Failed(100, "FILE_SYSTEM_UPLOAD_API_ERROR", err.Error())
+	}
+	return OK(nil)
+}
+
+// CQUploadPrivateFile 扩展API-上传私聊文件
+//
+// @route(upload_private_file)
+func (bot *CQBot) CQUploadPrivateFile(userID int64, file, name string) global.MSG {
+	target := message.Source{
+		SourceType: message.SourcePrivate,
+		PrimaryID:  userID,
+	}
+	fileBody, err := os.Open(file)
+	if err != nil {
+		log.Warnf("上传私聊文件 %v 失败: %+v", file, err)
+		return Failed(100, "OPEN_FILE_ERROR", "打开文件失败")
+	}
+	defer func() { _ = fileBody.Close() }()
+	localFile := &client.LocalFile{
+		FileName: name,
+		Body:     fileBody,
+	}
+	if err := bot.Client.UploadFile(target, localFile); err != nil {
+		log.Warnf("上传私聊 %v 文件 %v 失败: %+v", userID, file, err)
 		return Failed(100, "FILE_SYSTEM_UPLOAD_API_ERROR", err.Error())
 	}
 	return OK(nil)
@@ -534,11 +635,11 @@ func (bot *CQBot) CQUploadGroupFile(groupID int64, file, name, folder string) gl
 func (bot *CQBot) CQGroupFileCreateFolder(groupID int64, parentID, name string) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	if err = fs.CreateFolder(parentID, name); err != nil {
-		log.Errorf("创建群 %v 文件夹失败: %v", groupID, err)
+		log.Warnf("创建群 %v 文件夹失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	return OK(nil)
@@ -551,11 +652,11 @@ func (bot *CQBot) CQGroupFileCreateFolder(groupID int64, parentID, name string) 
 func (bot *CQBot) CQGroupFileDeleteFolder(groupID int64, id string) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	if err = fs.DeleteFolder(id); err != nil {
-		log.Errorf("删除群 %v 文件夹 %v 时出现文件: %v", groupID, id, err)
+		log.Warnf("删除群 %v 文件夹 %v 时出现文件: %v", groupID, id, err)
 		return Failed(200, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	return OK(nil)
@@ -564,15 +665,15 @@ func (bot *CQBot) CQGroupFileDeleteFolder(groupID int64, id string) global.MSG {
 // CQGroupFileDeleteFile 拓展API-删除群文件
 //
 // @route(delete_group_file)
-// @rename(id->file_id)
+// @rename(id->file_id, bus_id->"[busid\x2Cbus_id].0")
 func (bot *CQBot) CQGroupFileDeleteFile(groupID int64, id string, busID int32) global.MSG {
 	fs, err := bot.Client.GetGroupFileSystem(groupID)
 	if err != nil {
-		log.Errorf("获取群 %v 文件系统信息失败: %v", groupID, err)
+		log.Warnf("获取群 %v 文件系统信息失败: %v", groupID, err)
 		return Failed(100, "FILE_SYSTEM_API_ERROR", err.Error())
 	}
 	if res := fs.DeleteFile("", id, busID); res != "" {
-		log.Errorf("删除群 %v 文件 %v 时出现文件: %v", groupID, id, res)
+		log.Warnf("删除群 %v 文件 %v 时出现文件: %v", groupID, id, res)
 		return Failed(200, "FILE_SYSTEM_API_ERROR", res)
 	}
 	return OK(nil)
@@ -595,7 +696,7 @@ func (bot *CQBot) CQGetWordSlices(content string) global.MSG {
 
 // CQSendMessage 发送消息
 //
-// @route(send_msg)
+// @route11(send_msg)
 // @rename(m->message)
 func (bot *CQBot) CQSendMessage(groupID, userID int64, m gjson.Result, messageType string, autoEscape bool) global.MSG {
 	switch {
@@ -611,10 +712,28 @@ func (bot *CQBot) CQSendMessage(groupID, userID int64, m gjson.Result, messageTy
 	return global.MSG{}
 }
 
+// CQSendForwardMessage 发送合并转发消息
+//
+// @route11(send_forward_msg)
+// @rename(m->messages)
+func (bot *CQBot) CQSendForwardMessage(groupID, userID int64, m gjson.Result, messageType string) global.MSG {
+	switch {
+	case messageType == "group":
+		return bot.CQSendGroupForwardMessage(groupID, m)
+	case messageType == "private":
+		fallthrough
+	case userID != 0:
+		return bot.CQSendPrivateForwardMessage(userID, m)
+	case groupID != 0:
+		return bot.CQSendGroupForwardMessage(groupID, m)
+	}
+	return global.MSG{}
+}
+
 // CQSendGroupMessage 发送群消息
 //
 // https://git.io/Jtz1c
-// @route(send_group_msg)
+// @route11(send_group_msg)
 // @rename(m->message)
 func (bot *CQBot) CQSendGroupMessage(groupID int64, m gjson.Result, autoEscape bool) global.MSG {
 	group := bot.Client.FindGroup(groupID)
@@ -636,23 +755,23 @@ func (bot *CQBot) CQSendGroupMessage(groupID int64, m gjson.Result, autoEscape b
 
 	var elem []message.IMessageElement
 	if m.Type == gjson.JSON {
-		elem = bot.ConvertObjectMessage(m, MessageSourceGroup)
+		elem = bot.ConvertObjectMessage(onebot.V11, m, message.SourceGroup)
 	} else {
 		str := m.String()
 		if str == "" {
-			log.Warn("群消息发送失败: 信息为空.")
+			log.Warnf("群 %v 消息发送失败: 信息为空.", groupID)
 			return Failed(100, "EMPTY_MSG_ERROR", "消息为空")
 		}
 		if autoEscape {
 			elem = []message.IMessageElement{message.NewText(str)}
 		} else {
-			elem = bot.ConvertStringMessage(str, MessageSourceGroup)
+			elem = bot.ConvertStringMessage(onebot.V11, str, message.SourceGroup)
 		}
 	}
 	fixAt(elem)
-	mid := bot.SendGroupMessage(groupID, &message.SendingMessage{Elements: elem})
-	if mid == -1 {
-		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
+	mid, err := bot.SendGroupMessage(groupID, &message.SendingMessage{Elements: elem})
+	if err != nil {
+		return Failed(100, "SEND_MSG_API_ERROR", err.Error())
 	}
 	log.Infof("发送群 %v(%v) 的消息: %v (%v)", group.Name, groupID, limitedString(m.String()), mid)
 	return OK(global.MSG{"message_id": mid})
@@ -678,7 +797,7 @@ func (bot *CQBot) CQSendGuildChannelMessage(guildID, channelID uint64, m gjson.R
 	fixAt := func(elem []message.IMessageElement) {
 		for _, e := range elem {
 			if at, ok := e.(*message.AtElement); ok && at.Target != 0 && at.Display == "" {
-				mem, _ := bot.Client.GuildService.GetGuildMemberProfileInfo(guildID, uint64(at.Target))
+				mem, _ := bot.Client.GuildService.FetchGuildMemberProfileInfo(guildID, uint64(at.Target))
 				if mem != nil {
 					at.Display = "@" + mem.Nickname
 				} else {
@@ -690,7 +809,7 @@ func (bot *CQBot) CQSendGuildChannelMessage(guildID, channelID uint64, m gjson.R
 
 	var elem []message.IMessageElement
 	if m.Type == gjson.JSON {
-		elem = bot.ConvertObjectMessage(m, MessageSourceGuildChannel)
+		elem = bot.ConvertObjectMessage(onebot.V11, m, message.SourceGuildChannel)
 	} else {
 		str := m.String()
 		if str == "" {
@@ -700,7 +819,7 @@ func (bot *CQBot) CQSendGuildChannelMessage(guildID, channelID uint64, m gjson.R
 		if autoEscape {
 			elem = []message.IMessageElement{message.NewText(str)}
 		} else {
-			elem = bot.ConvertStringMessage(str, MessageSourceGuildChannel)
+			elem = bot.ConvertStringMessage(onebot.V11, str, message.SourceGuildChannel)
 		}
 	}
 	fixAt(elem)
@@ -712,148 +831,201 @@ func (bot *CQBot) CQSendGuildChannelMessage(guildID, channelID uint64, m gjson.R
 	return OK(global.MSG{"message_id": mid})
 }
 
+func (bot *CQBot) uploadForwardElement(m gjson.Result, target int64, sourceType message.SourceType) *message.ForwardElement {
+	ts := time.Now().Add(-time.Minute * 5)
+	groupID := target
+	source := message.Source{SourceType: sourceType, PrimaryID: target}
+	if sourceType == message.SourcePrivate {
+		// ios 设备的合并转发来源群号不能为 0
+		if len(bot.Client.GroupList) == 0 {
+			groupID = 1
+		} else {
+			groupID = bot.Client.GroupList[0].Uin
+		}
+	}
+	builder := bot.Client.NewForwardMessageBuilder(groupID)
+
+	var convertMessage func(m gjson.Result) *message.ForwardMessage
+	convertMessage = func(m gjson.Result) *message.ForwardMessage {
+		fm := message.NewForwardMessage()
+		var w worker
+		resolveElement := func(elems []message.IMessageElement) []message.IMessageElement {
+			for i, elem := range elems {
+				p := &elems[i]
+				switch o := elem.(type) {
+				case *msg.LocalVideo:
+					w.do(func() {
+						gm, err := bot.uploadLocalVideo(source, o)
+						if err != nil {
+							log.Warnf(uploadFailedTemplate, "合并转发", target, "视频", err)
+						} else {
+							*p = gm
+						}
+					})
+				case *msg.LocalImage:
+					w.do(func() {
+						gm, err := bot.uploadLocalImage(source, o)
+						if err != nil {
+							log.Warnf(uploadFailedTemplate, "合并转发", target, "图片", err)
+						} else {
+							*p = gm
+						}
+					})
+				}
+			}
+			return elems
+		}
+
+		convert := func(e gjson.Result) *message.ForwardNode {
+			if e.Get("type").Str != "node" {
+				return nil
+			}
+			if e.Get("data.id").Exists() {
+				i := e.Get("data.id").Int()
+				m, _ := db.GetMessageByGlobalID(int32(i))
+				if m != nil {
+					mSource := message.SourcePrivate
+					if m.GetType() == "group" {
+						mSource = message.SourceGroup
+					}
+					msgTime := m.GetAttribute().Timestamp
+					if msgTime == 0 {
+						msgTime = ts.Unix()
+					}
+					return &message.ForwardNode{
+						SenderId:   m.GetAttribute().SenderUin,
+						SenderName: m.GetAttribute().SenderName,
+						Time:       int32(msgTime),
+						Message:    resolveElement(bot.ConvertContentMessage(m.GetContent(), mSource, false)),
+					}
+				}
+				log.Warnf("警告: 引用消息 %v 错误或数据库未开启.", e.Get("data.id").Str)
+				return nil
+			}
+			uin := e.Get("data.[user_id,uin].0").Int()
+			msgTime := e.Get("data.time").Int()
+			if msgTime == 0 {
+				msgTime = ts.Unix()
+			}
+			name := e.Get("data.[name,nickname].0").Str
+			c := e.Get("data.content")
+			if c.IsArray() {
+				nested := false
+				c.ForEach(func(_, value gjson.Result) bool {
+					if value.Get("type").Str == "node" {
+						nested = true
+						return false
+					}
+					return true
+				})
+				if nested { // 处理嵌套
+					nestedNode := builder.NestedNode()
+					builder.Link(nestedNode, convertMessage(c))
+					return &message.ForwardNode{
+						SenderId:   uin,
+						SenderName: name,
+						Time:       int32(msgTime),
+						Message:    []message.IMessageElement{nestedNode},
+					}
+				}
+			}
+			content := bot.ConvertObjectMessage(onebot.V11, c, sourceType)
+			if uin != 0 && name != "" && len(content) > 0 {
+				return &message.ForwardNode{
+					SenderId:   uin,
+					SenderName: name,
+					Time:       int32(msgTime),
+					Message:    resolveElement(content),
+				}
+			}
+			log.Warnf("警告: 非法 Forward node 将跳过. uin: %v name: %v content count: %v", uin, name, len(content))
+			return nil
+		}
+
+		if m.IsArray() {
+			for _, item := range m.Array() {
+				node := convert(item)
+				if node != nil {
+					fm.AddNode(node)
+				}
+			}
+		} else {
+			node := convert(m)
+			if node != nil {
+				fm.AddNode(node)
+			}
+		}
+
+		w.wait()
+		return fm
+	}
+	return builder.Main(convertMessage(m))
+}
+
 // CQSendGroupForwardMessage 扩展API-发送合并转发(群)
 //
 // https://docs.go-cqhttp.org/api/#%E5%8F%91%E9%80%81%E5%90%88%E5%B9%B6%E8%BD%AC%E5%8F%91-%E7%BE%A4
-// @route(send_group_forward_msg)
+// @route11(send_group_forward_msg)
 // @rename(m->messages)
 func (bot *CQBot) CQSendGroupForwardMessage(groupID int64, m gjson.Result) global.MSG {
 	if m.Type != gjson.JSON {
 		return Failed(100)
 	}
-	fm := message.NewForwardMessage()
-	ts := time.Now().Add(-time.Minute * 5)
-	hasCustom := false
-	m.ForEach(func(_, item gjson.Result) bool {
-		if item.Get("data.uin").Exists() || item.Get("data.user_id").Exists() {
-			hasCustom = true
-			return false
-		}
-		return true
+	source := message.Source{
+		SourceType: message.SourcePrivate,
+		PrimaryID:  0,
+	}
+	fe := bot.uploadForwardElement(m, groupID, message.SourceGroup)
+	if fe == nil {
+		return Failed(100, "EMPTY_NODES", "未找到任何可发送的合并转发信息")
+	}
+	ret := bot.Client.SendGroupForwardMessage(groupID, fe)
+	if ret == nil || ret.Id == -1 {
+		log.Warnf("合并转发(群 %v)消息发送失败: 账号可能被风控.", groupID)
+		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
+	}
+	mid := bot.InsertGroupMessage(ret, source)
+	log.Infof("发送群 %v(%v)  的合并转发消息: %v (%v)", groupID, groupID, limitedString(m.String()), mid)
+	return OK(global.MSG{
+		"message_id": mid,
+		"forward_id": fe.ResId,
 	})
+}
 
-	resolveElement := func(elems []message.IMessageElement) []message.IMessageElement {
-		for i, elem := range elems {
-			switch elem.(type) {
-			case *LocalImageElement, *LocalVideoElement:
-				gm, err := bot.uploadMedia(elem, groupID, true)
-				if err != nil {
-					log.Warnf("警告: 群 %d %s上传失败: %v", groupID, elem.Type().String(), err)
-					continue
-				}
-				elems[i] = gm
-			}
-		}
-		return elems
+// CQSendPrivateForwardMessage 扩展API-发送合并转发(好友)
+//
+// https://docs.go-cqhttp.org/api/#%E5%8F%91%E9%80%81%E5%90%88%E5%B9%B6%E8%BD%AC%E5%8F%91-%E7%BE%A4
+// @route11(send_private_forward_msg)
+// @rename(m->messages)
+func (bot *CQBot) CQSendPrivateForwardMessage(userID int64, m gjson.Result) global.MSG {
+	if m.Type != gjson.JSON {
+		return Failed(100)
 	}
-
-	var convert func(e gjson.Result) *message.ForwardNode
-	convert = func(e gjson.Result) *message.ForwardNode {
-		if e.Get("type").Str != "node" {
-			return nil
-		}
-		ts.Add(time.Second)
-		if e.Get("data.id").Exists() {
-			i := e.Get("data.id").Int()
-			m, _ := db.GetGroupMessageByGlobalID(int32(i))
-			if m != nil {
-				return &message.ForwardNode{
-					SenderId:   m.Attribute.SenderUin,
-					SenderName: m.Attribute.SenderName,
-					Time: func() int32 {
-						msgTime := m.Attribute.Timestamp
-						if hasCustom && msgTime == 0 {
-							return int32(ts.Unix())
-						}
-						return int32(msgTime)
-					}(),
-					Message: resolveElement(bot.ConvertContentMessage(m.Content, MessageSourceGroup)),
-				}
-			}
-			log.Warnf("警告: 引用消息 %v 错误或数据库未开启.", e.Get("data.id").Str)
-			return nil
-		}
-		uin := e.Get("data.[user_id,uin].0").Int()
-		msgTime := e.Get("data.time").Int()
-		if msgTime == 0 {
-			msgTime = ts.Unix()
-		}
-		name := e.Get("data.name").Str
-		c := e.Get("data.content")
-		if c.IsArray() {
-			nested := false
-			c.ForEach(func(_, value gjson.Result) bool {
-				if value.Get("type").Str == "node" {
-					nested = true
-					return false
-				}
-				return true
-			})
-			if nested { // 处理嵌套
-				nest := message.NewForwardMessage()
-				for _, item := range c.Array() {
-					node := convert(item)
-					if node != nil {
-						nest.AddNode(node)
-					}
-				}
-				elem := bot.Client.UploadGroupForwardMessage(groupID, nest)
-				return &message.ForwardNode{
-					SenderId:   uin,
-					SenderName: name,
-					Time:       int32(msgTime),
-					Message:    []message.IMessageElement{elem},
-				}
-			}
-		}
-		content := bot.ConvertObjectMessage(e.Get("data.content"), MessageSourceGroup)
-		if uin != 0 && name != "" && len(content) > 0 {
-			return &message.ForwardNode{
-				SenderId:   uin,
-				SenderName: name,
-				Time:       int32(msgTime),
-				Message:    resolveElement(content),
-			}
-		}
-		log.Warnf("警告: 非法 Forward node 将跳过. uin: %v name: %v content count: %v", uin, name, len(content))
-		return nil
+	fe := bot.uploadForwardElement(m, userID, message.SourcePrivate)
+	if fe == nil {
+		return Failed(100, "EMPTY_NODES", "未找到任何可发送的合并转发信息")
 	}
-	if m.IsArray() {
-		for _, item := range m.Array() {
-			node := convert(item)
-			if node != nil {
-				fm.AddNode(node)
-			}
-		}
-	} else {
-		node := convert(m)
-		if node != nil {
-			fm.AddNode(node)
-		}
+	mid := bot.SendPrivateMessage(userID, 0, &message.SendingMessage{Elements: []message.IMessageElement{fe}})
+	if mid == -1 {
+		log.Warnf("合并转发(好友 %v)消息发送失败: 账号可能被风控.", userID)
+		return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
 	}
-	if fm.Length() > 0 {
-		fe := bot.Client.UploadGroupForwardMessage(groupID, fm)
-		ret := bot.Client.SendGroupForwardMessage(groupID, fe)
-		if ret == nil || ret.Id == -1 {
-			log.Warnf("合并转发(群)消息发送失败: 账号可能被风控.")
-			return Failed(100, "SEND_MSG_API_ERROR", "请参考 go-cqhttp 端输出")
-		}
-		return OK(global.MSG{
-			"message_id": bot.InsertGroupMessage(ret),
-		})
-	}
-	return Failed(100, "EMPTY_NODES", "未找到任何可发送的合并转发信息")
+	log.Infof("发送好友 %v(%v)  的合并转发消息: %v (%v)", userID, userID, limitedString(m.String()), mid)
+	return OK(global.MSG{
+		"message_id": mid,
+		"forward_id": fe.ResId,
+	})
 }
 
 // CQSendPrivateMessage 发送私聊消息
 //
 // https://git.io/Jtz1l
-// @route(send_private_msg)
+// @route11(send_private_msg)
 // @rename(m->message)
 func (bot *CQBot) CQSendPrivateMessage(userID int64, groupID int64, m gjson.Result, autoEscape bool) global.MSG {
 	var elem []message.IMessageElement
 	if m.Type == gjson.JSON {
-		elem = bot.ConvertObjectMessage(m, MessageSourcePrivate)
+		elem = bot.ConvertObjectMessage(onebot.V11, m, message.SourcePrivate)
 	} else {
 		str := m.String()
 		if str == "" {
@@ -862,7 +1034,7 @@ func (bot *CQBot) CQSendPrivateMessage(userID int64, groupID int64, m gjson.Resu
 		if autoEscape {
 			elem = []message.IMessageElement{message.NewText(str)}
 		} else {
-			elem = bot.ConvertStringMessage(str, MessageSourcePrivate)
+			elem = bot.ConvertStringMessage(onebot.V11, str, message.SourcePrivate)
 		}
 	}
 	mid := bot.SendPrivateMessage(userID, groupID, &message.SendingMessage{Elements: elem})
@@ -915,6 +1087,17 @@ func (bot *CQBot) CQSetGroupName(groupID int64, name string) global.MSG {
 	return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
 }
 
+// CQGetGroupMemo 扩展API-获取群公告
+// @route(_get_group_notice)
+func (bot *CQBot) CQGetGroupMemo(groupID int64) global.MSG {
+	r, err := bot.Client.GetGroupNotice(groupID)
+	if err != nil {
+		return Failed(100, "获取群公告失败", err.Error())
+	}
+
+	return OK(r)
+}
+
 // CQSetGroupMemo 扩展API-发送群公告
 //
 // https://docs.go-cqhttp.org/api/#%E5%8F%91%E9%80%81%E7%BE%A4%E5%85%AC%E5%91%8A
@@ -930,15 +1113,32 @@ func (bot *CQBot) CQSetGroupMemo(groupID int64, msg, img string) global.MSG {
 			if err != nil {
 				return Failed(100, "IMAGE_NOT_FOUND", "图片未找到")
 			}
-			err = bot.Client.AddGroupNoticeWithPic(groupID, msg, data)
+			noticeID, err := bot.Client.AddGroupNoticeWithPic(groupID, msg, data)
 			if err != nil {
 				return Failed(100, "SEND_NOTICE_ERROR", err.Error())
 			}
-		} else {
-			err := bot.Client.AddGroupNoticeSimple(groupID, msg)
-			if err != nil {
-				return Failed(100, "SEND_NOTICE_ERROR", err.Error())
-			}
+			return OK(global.MSG{"notice_id": noticeID})
+		}
+		noticeID, err := bot.Client.AddGroupNoticeSimple(groupID, msg)
+		if err != nil {
+			return Failed(100, "SEND_NOTICE_ERROR", err.Error())
+		}
+		return OK(global.MSG{"notice_id": noticeID})
+	}
+	return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
+}
+
+// CQDelGroupMemo 扩展API-删除群公告
+// @route(_del_group_notice)
+// @rename(fid->notice_id)
+func (bot *CQBot) CQDelGroupMemo(groupID int64, fid string) global.MSG {
+	if g := bot.Client.FindGroup(groupID); g != nil {
+		if g.SelfPermission() == client.Member {
+			return Failed(100, "PERMISSION_DENIED", "权限不足")
+		}
+		err := bot.Client.DelGroupNotice(groupID, fid)
+		if err != nil {
+			return Failed(100, "DELETE_NOTICE_ERROR", err.Error())
 		}
 		return OK(nil)
 	}
@@ -952,13 +1152,15 @@ func (bot *CQBot) CQSetGroupMemo(groupID int64, msg, img string) global.MSG {
 // @rename(msg->message, block->reject_add_request)
 func (bot *CQBot) CQSetGroupKick(groupID int64, userID int64, msg string, block bool) global.MSG {
 	if g := bot.Client.FindGroup(groupID); g != nil {
-		if m := g.FindMember(userID); m != nil {
-			err := m.Kick(msg, block)
-			if err != nil {
-				return Failed(100, "NOT_MANAGEABLE", "机器人权限不足")
-			}
-			return OK(nil)
+		m := g.FindMember(userID)
+		if m == nil {
+			return Failed(100, "MEMBER_NOT_FOUND", "人员不存在")
 		}
+		err := m.Kick(msg, block)
+		if err != nil {
+			return Failed(100, "NOT_MANAGEABLE", "机器人权限不足")
+		}
+		return OK(nil)
 	}
 	return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
 }
@@ -973,7 +1175,7 @@ func (bot *CQBot) CQSetGroupBan(groupID, userID int64, duration uint32) global.M
 		if m := g.FindMember(userID); m != nil {
 			err := m.Mute(duration)
 			if err != nil {
-				if duration > 2592000 {
+				if duration >= 2592000 {
 					return Failed(100, "DURATION_IS_NOT_IN_RANGE", "非法的禁言时长")
 				}
 				return Failed(100, "NOT_MANAGEABLE", "机器人权限不足")
@@ -1035,9 +1237,9 @@ func (bot *CQBot) CQProcessFriendRequest(flag string, approve bool) global.MSG {
 		return Failed(100, "FLAG_NOT_FOUND", "FLAG不存在")
 	}
 	if approve {
-		req.(*client.NewFriendRequest).Accept()
+		req.Accept()
 	} else {
-		req.(*client.NewFriendRequest).Reject()
+		req.Reject()
 	}
 	return OK(nil)
 }
@@ -1051,14 +1253,14 @@ func (bot *CQBot) CQProcessFriendRequest(flag string, approve bool) global.MSG {
 func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bool) global.MSG {
 	msgs, err := bot.Client.GetGroupSystemMessages()
 	if err != nil {
-		log.Errorf("获取群系统消息失败: %v", err)
+		log.Warnf("获取群系统消息失败: %v", err)
 		return Failed(100, "SYSTEM_MSG_API_ERROR", err.Error())
 	}
 	if subType == "add" {
 		for _, req := range msgs.JoinRequests {
 			if strconv.FormatInt(req.RequestId, 10) == flag {
 				if req.Checked {
-					log.Errorf("处理群系统消息失败: 无法操作已处理的消息.")
+					log.Warnf("处理群系统消息失败: 无法操作已处理的消息.")
 					return Failed(100, "FLAG_HAS_BEEN_CHECKED", "消息已被处理")
 				}
 				if approve {
@@ -1073,7 +1275,7 @@ func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bo
 		for _, req := range msgs.InvitedRequests {
 			if strconv.FormatInt(req.RequestId, 10) == flag {
 				if req.Checked {
-					log.Errorf("处理群系统消息失败: 无法操作已处理的消息.")
+					log.Warnf("处理群系统消息失败: 无法操作已处理的消息.")
 					return Failed(100, "FLAG_HAS_BEEN_CHECKED", "消息已被处理")
 				}
 				if approve {
@@ -1085,7 +1287,7 @@ func (bot *CQBot) CQProcessGroupRequest(flag, subType, reason string, approve bo
 			}
 		}
 	}
-	log.Errorf("处理群系统消息失败: 消息 %v 不存在.", flag)
+	log.Warnf("处理群系统消息失败: 消息 %v 不存在.", flag)
 	return Failed(100, "FLAG_NOT_FOUND", "FLAG不存在")
 }
 
@@ -1144,25 +1346,17 @@ func (bot *CQBot) CQSetGroupAdmin(groupID, userID int64, enable bool) global.MSG
 	return OK(nil)
 }
 
-// CQGetVipInfo 扩展API-获取VIP信息
+// CQSetGroupAnonymous 群组匿名
 //
-// https://docs.go-cqhttp.org/api/#%E8%8E%B7%E5%8F%96vip%E4%BF%A1%E6%81%AF
-// @route(_get_vip_info)
-func (bot *CQBot) CQGetVipInfo(userID int64) global.MSG {
-	vip, err := bot.Client.GetVipInfo(userID)
-	if err != nil {
-		return Failed(100, "VIP_API_ERROR", err.Error())
+// https://beautyyu.one
+// @route(set_group_anonymous)
+// @default(enable=true)
+func (bot *CQBot) CQSetGroupAnonymous(groupID int64, enable bool) global.MSG {
+	if g := bot.Client.FindGroup(groupID); g != nil {
+		g.SetAnonymous(enable)
+		return OK(nil)
 	}
-	msg := global.MSG{
-		"user_id":          vip.Uin,
-		"nickname":         vip.Name,
-		"level":            vip.Level,
-		"level_speed":      vip.LevelSpeed,
-		"vip_level":        vip.VipLevel,
-		"vip_growth_speed": vip.VipGrowthSpeed,
-		"vip_growth_total": vip.VipGrowthTotal,
-	}
-	return OK(msg)
+	return Failed(100, "GROUP_NOT_FOUND", "群聊不存在")
 }
 
 // CQGetGroupHonorInfo 获取群荣誉信息
@@ -1194,40 +1388,50 @@ func (bot *CQBot) CQGetGroupHonorInfo(groupID int64, t string) global.MSG {
 				}
 			}
 			msg["talkative_list"] = convertMem(honor.TalkativeList)
+		} else {
+			log.Infof("获取群龙王出错：%v", err)
 		}
 	}
 
 	if t == "performer" || t == "all" {
 		if honor, err := bot.Client.GetGroupHonorInfo(groupID, client.Performer); err == nil {
-			msg["performer_lis"] = convertMem(honor.ActorList)
+			msg["performer_list"] = convertMem(honor.ActorList)
+		} else {
+			log.Infof("获取群聊之火出错：%v", err)
 		}
 	}
 
 	if t == "legend" || t == "all" {
 		if honor, err := bot.Client.GetGroupHonorInfo(groupID, client.Legend); err == nil {
 			msg["legend_list"] = convertMem(honor.LegendList)
+		} else {
+			log.Infof("获取群聊炽焰出错：%v", err)
 		}
 	}
 
 	if t == "strong_newbie" || t == "all" {
 		if honor, err := bot.Client.GetGroupHonorInfo(groupID, client.StrongNewbie); err == nil {
 			msg["strong_newbie_list"] = convertMem(honor.StrongNewbieList)
+		} else {
+			log.Infof("获取冒尖小春笋出错：%v", err)
 		}
 	}
 
 	if t == "emotion" || t == "all" {
 		if honor, err := bot.Client.GetGroupHonorInfo(groupID, client.Emotion); err == nil {
 			msg["emotion_list"] = convertMem(honor.EmotionList)
+		} else {
+			log.Infof("获取快乐之源出错：%v", err)
 		}
 	}
-
 	return OK(msg)
 }
 
 // CQGetStrangerInfo 获取陌生人信息
 //
 // https://git.io/Jtz17
-// @route(get_stranger_info)
+// @route11(get_stranger_info)
+// @route12(get_user_info)
 func (bot *CQBot) CQGetStrangerInfo(userID int64) global.MSG {
 	info, err := bot.Client.GetSummaryInfo(userID)
 	if err != nil {
@@ -1246,16 +1450,18 @@ func (bot *CQBot) CQGetStrangerInfo(userID int64) global.MSG {
 			// unknown = 0x2
 			return "unknown"
 		}(),
+		"sign":       info.Sign,
 		"age":        info.Age,
 		"level":      info.Level,
 		"login_days": info.LoginDays,
+		"vip_level":  info.VipLevel,
 	})
 }
 
 // CQHandleQuickOperation 隐藏API-对事件执行快速操作
 //
 // https://git.io/Jtz15
-// @route(".handle_quick_operation")
+// @route11(".handle_quick_operation")
 func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) global.MSG {
 	postType := context.Get("post_type").Str
 
@@ -1268,7 +1474,7 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) global
 
 		if reply.Exists() {
 			autoEscape := param.EnsureBool(operation.Get("auto_escape"), false)
-			at := operation.Get("at_sender").Bool() && !isAnonymous && msgType == "group"
+			at := !isAnonymous && operation.Get("at_sender").Bool() && msgType == "group"
 			if at && reply.IsArray() {
 				// 在 reply 数组头部插入CQ码
 				replySegments := make([]global.MSG, 0)
@@ -1314,7 +1520,7 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) global
 			if operation.Get("delete").Bool() {
 				bot.CQDeleteMessage(int32(context.Get("message_id").Int()))
 			}
-			if operation.Get("kick").Bool() && !isAnonymous {
+			if !isAnonymous && operation.Get("kick").Bool() {
 				bot.CQSetGroupKick(context.Get("group_id").Int(), context.Get("user_id").Int(), "", operation.Get("reject_add_request").Bool())
 			}
 			if operation.Get("ban").Bool() {
@@ -1349,7 +1555,7 @@ func (bot *CQBot) CQHandleQuickOperation(context, operation gjson.Result) global
 func (bot *CQBot) CQGetImage(file string) global.MSG {
 	var b []byte
 	var err error
-	if cache.EnableCacheDB && strings.HasSuffix(file, ".image") {
+	if strings.HasSuffix(file, ".image") {
 		var f []byte
 		f, err = hex.DecodeString(strings.TrimSuffix(file, ".image"))
 		b = cache.Image.Get(f)
@@ -1372,12 +1578,8 @@ func (bot *CQBot) CQGetImage(file string) global.MSG {
 		}
 		local := path.Join(global.CachePath, file+path.Ext(msg["filename"].(string)))
 		if !global.PathExists(local) {
-			if body, err := global.HTTPGetReadCloser(msg["url"].(string)); err == nil {
-				f, _ := os.OpenFile(local, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0o0644)
-				_, _ = f.ReadFrom(body)
-				_ = body.Close()
-				_ = f.Close()
-			} else {
+			r := download.Request{URL: msg["url"].(string)}
+			if err := r.WriteToFile(local); err != nil {
 				log.Warnf("下载图片 %v 时出现错误: %v", msg["url"], err)
 				return Failed(100, "DOWNLOAD_IMAGE_ERROR", err.Error())
 			}
@@ -1396,18 +1598,18 @@ func (bot *CQBot) CQDownloadFile(url string, headers gjson.Result, threadCount i
 	h := map[string]string{}
 	if headers.IsArray() {
 		for _, sub := range headers.Array() {
-			str := strings.SplitN(sub.String(), "=", 2)
-			if len(str) == 2 {
-				h[str[0]] = str[1]
+			first, second, ok := strings.Cut(sub.String(), "=")
+			if ok {
+				h[first] = second
 			}
 		}
 	}
 	if headers.Type == gjson.String {
 		lines := strings.Split(headers.String(), "\r\n")
 		for _, sub := range lines {
-			str := strings.SplitN(sub, "=", 2)
-			if len(str) == 2 {
-				h[str[0]] = str[1]
+			first, second, ok := strings.Cut(sub, "=")
+			if ok {
+				h[first] = second
 			}
 		}
 	}
@@ -1420,7 +1622,8 @@ func (bot *CQBot) CQDownloadFile(url string, headers gjson.Result, threadCount i
 			return Failed(100, "DELETE_FILE_ERROR", err.Error())
 		}
 	}
-	if err := global.DownloadFileMultiThreading(url, file, 0, threadCount, h); err != nil {
+	r := download.Request{URL: url, Header: h}
+	if err := r.WriteToFileMultiThreading(file, threadCount); err != nil {
 		log.Warnf("下载链接 %v 时出现错误: %v", url, err)
 		return Failed(100, "DOWNLOAD_FILE_ERROR", err.Error())
 	}
@@ -1440,20 +1643,32 @@ func (bot *CQBot) CQGetForwardMessage(resID string) global.MSG {
 	if m == nil {
 		return Failed(100, "MSG_NOT_FOUND", "消息不存在")
 	}
-	r := make([]global.MSG, 0, len(m.Nodes))
-	for _, n := range m.Nodes {
-		bot.checkMedia(n.Message, 0)
-		r = append(r, global.MSG{
-			"sender": global.MSG{
-				"user_id":  n.SenderId,
-				"nickname": n.SenderName,
-			},
-			"time":    n.Time,
-			"content": ToFormattedMessage(n.Message, MessageSource{SourceType: MessageSourceGroup}, false),
-		})
+
+	var transformNodes func(nodes []*message.ForwardNode) []global.MSG
+	transformNodes = func(nodes []*message.ForwardNode) []global.MSG {
+		r := make([]global.MSG, len(nodes))
+		for i, n := range nodes {
+			bot.checkMedia(n.Message, 0)
+			content := ToFormattedMessage(n.Message, message.Source{SourceType: message.SourceGroup})
+			if len(n.Message) == 1 {
+				if forward, ok := n.Message[0].(*message.ForwardMessage); ok {
+					content = transformNodes(forward.Nodes)
+				}
+			}
+			r[i] = global.MSG{
+				"sender": global.MSG{
+					"user_id":  n.SenderId,
+					"nickname": n.SenderName,
+				},
+				"time":     n.Time,
+				"content":  content,
+				"group_id": n.GroupId,
+			}
+		}
+		return r
 	}
 	return OK(global.MSG{
-		"messages": r,
+		"messages": transformNodes(m.Nodes),
 	})
 }
 
@@ -1483,9 +1698,73 @@ func (bot *CQBot) CQGetMessage(messageID int32) global.MSG {
 	switch o := msg.(type) {
 	case *db.StoredGroupMessage:
 		m["group_id"] = o.GroupCode
-		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, MessageSourceGroup), MessageSource{SourceType: MessageSourceGroup, PrimaryID: uint64(o.GroupCode)}, false)
+		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, message.SourceGroup, false), message.Source{SourceType: message.SourceGroup, PrimaryID: o.GroupCode})
 	case *db.StoredPrivateMessage:
-		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, MessageSourcePrivate), MessageSource{SourceType: MessageSourcePrivate}, false)
+		m["message"] = ToFormattedMessage(bot.ConvertContentMessage(o.Content, message.SourcePrivate, false), message.Source{SourceType: message.SourcePrivate})
+	}
+	return OK(m)
+}
+
+// CQGetGuildMessage 获取频道消息
+// @route(get_guild_msg)
+func (bot *CQBot) CQGetGuildMessage(messageID string, noCache bool) global.MSG {
+	source, seq := decodeGuildMessageID(messageID)
+	if source.SourceType == 0 {
+		log.Warnf("获取消息时出现错误: 无效消息ID")
+		return Failed(100, "INVALID_MESSAGE_ID", "无效消息ID")
+	}
+	m := global.MSG{
+		"message_id": messageID,
+		"message_source": func() string {
+			if source.SourceType == message.SourceGuildDirect {
+				return "direct"
+			}
+			return "channel"
+		}(),
+		"message_seq": seq,
+		"guild_id":    fU64(uint64(source.PrimaryID)),
+		"reactions":   []int{},
+	}
+	// nolint: exhaustive
+	switch source.SourceType {
+	case message.SourceGuildChannel:
+		m["channel_id"] = fU64(uint64(source.SecondaryID))
+		if noCache {
+			pull, err := bot.Client.GuildService.PullGuildChannelMessage(uint64(source.PrimaryID), uint64(source.SecondaryID), seq, seq)
+			if err != nil {
+				log.Warnf("获取消息时出现错误: %v", err)
+				return Failed(100, "API_ERROR", err.Error())
+			}
+			if len(m) == 0 {
+				log.Warnf("获取消息时出现错误: 消息不存在")
+				return Failed(100, "MSG_NOT_FOUND", "消息不存在")
+			}
+			m["time"] = pull[0].Time
+			m["sender"] = global.MSG{
+				"user_id":  pull[0].Sender.TinyId,
+				"tiny_id":  fU64(pull[0].Sender.TinyId),
+				"nickname": pull[0].Sender.Nickname,
+			}
+			m["message"] = ToFormattedMessage(pull[0].Elements, source)
+			m["reactions"] = convertReactions(pull[0].Reactions)
+			bot.InsertGuildChannelMessage(pull[0])
+		} else {
+			channelMsgByDB, err := db.GetGuildChannelMessageByID(messageID)
+			if err != nil {
+				log.Warnf("获取消息时出现错误: %v", err)
+				return Failed(100, "MSG_NOT_FOUND", "消息不存在")
+			}
+			m["time"] = channelMsgByDB.Attribute.Timestamp
+			m["sender"] = global.MSG{
+				"user_id":  channelMsgByDB.Attribute.SenderTinyID,
+				"tiny_id":  fU64(channelMsgByDB.Attribute.SenderTinyID),
+				"nickname": channelMsgByDB.Attribute.SenderName,
+			}
+			m["message"] = ToFormattedMessage(bot.ConvertContentMessage(channelMsgByDB.Content, message.SourceGuildChannel, false), source)
+		}
+	case message.SourceGuildDirect:
+		// todo(mrs4s): 支持 direct 消息
+		m["tiny_id"] = fU64(uint64(source.SecondaryID))
 	}
 	return OK(m)
 }
@@ -1524,12 +1803,16 @@ func (bot *CQBot) CQGetGroupMessageHistory(groupID int64, seq int64) global.MSG 
 		log.Warnf("获取群历史消息失败: %v", err)
 		return Failed(100, "MESSAGES_API_ERROR", err.Error())
 	}
-	ms := make([]global.MSG, 0, len(msg))
+	source := message.Source{
+		SourceType: message.SourcePrivate,
+		PrimaryID:  0,
+	}
+	ms := make([]*event, 0, len(msg))
 	for _, m := range msg {
 		bot.checkMedia(m.Elements, groupID)
-		id := bot.InsertGroupMessage(m)
+		id := bot.InsertGroupMessage(m, source)
 		t := bot.formatGroupMessage(m)
-		t["message_id"] = id
+		t.Others["message_id"] = id
 		ms = append(ms, t)
 	}
 	return OK(global.MSG{
@@ -1564,7 +1847,7 @@ func (bot *CQBot) CQGetOnlineClients(noCache bool) global.MSG {
 // CQCanSendImage 检查是否可以发送图片(此处永远返回true)
 //
 // https://git.io/Jtz1N
-// @route(can_send_image)
+// @route11(can_send_image)
 func (bot *CQBot) CQCanSendImage() global.MSG {
 	return OK(global.MSG{"yes": true})
 }
@@ -1572,7 +1855,7 @@ func (bot *CQBot) CQCanSendImage() global.MSG {
 // CQCanSendRecord 检查是否可以发送语音(此处永远返回true)
 //
 // https://git.io/Jtz1x
-// @route(can_send_record)
+// @route11(can_send_record)
 func (bot *CQBot) CQCanSendRecord() global.MSG {
 	return OK(global.MSG{"yes": true})
 }
@@ -1580,11 +1863,14 @@ func (bot *CQBot) CQCanSendRecord() global.MSG {
 // CQOcrImage 扩展API-图片OCR
 //
 // https://docs.go-cqhttp.org/api/#%E5%9B%BE%E7%89%87-ocr
-// @route(ocr_image)
-// @alias(.ocr_image)
+// @route(ocr_image,".ocr_image")
 // @rename(image_id->image)
 func (bot *CQBot) CQOcrImage(imageID string) global.MSG {
-	img, err := bot.makeImageOrVideoElem(map[string]string{"file": imageID}, false, MessageSourceGroup)
+	// TODO: fix this
+	var elem msg.Element
+	elem.Type = "image"
+	elem.Data = []msg.Pair{{K: "file", V: imageID}}
+	img, err := bot.makeImageOrVideoElem(elem, false, message.SourceGroup)
 	if err != nil {
 		log.Warnf("load image error: %v", err)
 		return Failed(100, "LOAD_FILE_ERROR", err.Error())
@@ -1624,12 +1910,10 @@ func (bot *CQBot) CQSetGroupAnonymousBan(groupID int64, flag string, duration in
 		return Failed(100, "INVALID_FLAG", "无效的flag")
 	}
 	if g := bot.Client.FindGroup(groupID); g != nil {
-		s := strings.SplitN(flag, "|", 2)
-		if len(s) != 2 {
+		id, nick, ok := strings.Cut(flag, "|")
+		if !ok {
 			return Failed(100, "INVALID_FLAG", "无效的flag")
 		}
-		id := s[0]
-		nick := s[1]
 		if err := g.MuteAnonymous(id, nick, duration); err != nil {
 			log.Warnf("anonymous ban error: %v", err)
 			return Failed(100, "CALL_API_ERROR", err.Error())
@@ -1643,15 +1927,22 @@ func (bot *CQBot) CQSetGroupAnonymousBan(groupID int64, flag string, duration in
 //
 // https://git.io/JtzMe
 // @route(get_status)
-func (bot *CQBot) CQGetStatus() global.MSG {
+func (bot *CQBot) CQGetStatus(spec *onebot.Spec) global.MSG {
+	if spec.Version == 11 {
+		return OK(global.MSG{
+			"app_initialized": true,
+			"app_enabled":     true,
+			"plugins_good":    nil,
+			"app_good":        true,
+			"online":          bot.Client.Online.Load(),
+			"good":            bot.Client.Online.Load(),
+			"stat":            bot.Client.GetStatistics(),
+		})
+	}
 	return OK(global.MSG{
-		"app_initialized": true,
-		"app_enabled":     true,
-		"plugins_good":    nil,
-		"app_good":        true,
-		"online":          bot.Client.Online,
-		"good":            bot.Client.Online,
-		"stat":            bot.Client.GetStatistics(),
+		"online": bot.Client.Online.Load(),
+		"good":   bot.Client.Online.Load(),
+		"stat":   bot.Client.GetStatistics(),
 	})
 }
 
@@ -1729,7 +2020,7 @@ func (bot *CQBot) CQCheckURLSafely(url string) global.MSG {
 // CQGetVersionInfo 获取版本信息
 //
 // https://git.io/JtwUs
-// @route(get_version_info)
+// @route11(get_version_info)
 func (bot *CQBot) CQGetVersionInfo() global.MSG {
 	wd, _ := os.Getwd()
 	return OK(global.MSG{
@@ -1739,29 +2030,14 @@ func (bot *CQBot) CQGetVersionInfo() global.MSG {
 		"protocol_version":           "v11",
 		"coolq_directory":            wd,
 		"coolq_edition":              "pro",
-		"go-cqhttp":                  true,
+		"go_cqhttp":                  true,
 		"plugin_version":             "4.15.0",
 		"plugin_build_number":        99,
 		"plugin_build_configuration": "release",
 		"runtime_version":            runtime.Version(),
 		"runtime_os":                 runtime.GOOS,
 		"version":                    base.Version,
-		"protocol": func() int {
-			switch client.SystemDeviceInfo.Protocol {
-			case client.IPad:
-				return 0
-			case client.AndroidPhone:
-				return 1
-			case client.AndroidWatch:
-				return 2
-			case client.MacOS:
-				return 3
-			case client.QiDian:
-				return 4
-			default:
-				return -1
-			}
-		}(),
+		"protocol_name":              bot.Client.Device().Protocol,
 	})
 }
 
@@ -1784,6 +2060,15 @@ func (bot *CQBot) CQGetModelShow(model string) global.MSG {
 	return OK(global.MSG{
 		"variants": a,
 	})
+}
+
+// CQSendGroupSign 群打卡
+//
+// https://club.vip.qq.com/onlinestatus/set
+// @route(send_group_sign)
+func (bot *CQBot) CQSendGroupSign(groupID int64) global.MSG {
+	bot.Client.SendGroupSign(groupID)
+	return OK(nil)
 }
 
 // CQSetModelShow 设置在线机型
@@ -1816,6 +2101,27 @@ func (bot *CQBot) CQMarkMessageAsRead(msgID int32) global.MSG {
 	return OK(nil)
 }
 
+// CQSetQQProfile 设置 QQ 资料
+//
+// @route(set_qq_profile)
+func (bot *CQBot) CQSetQQProfile(nickname, company, email, college, personalNote gjson.Result) global.MSG {
+	u := client.NewProfileDetailUpdate()
+
+	fi := func(f gjson.Result, do func(value string) client.ProfileDetailUpdate) {
+		if f.Exists() {
+			do(f.String())
+		}
+	}
+
+	fi(nickname, u.Nick)
+	fi(company, u.Company)
+	fi(email, u.Email)
+	fi(college, u.College)
+	fi(personalNote, u.PersonalNote)
+	bot.Client.UpdateProfile(u)
+	return OK(nil)
+}
+
 // CQReloadEventFilter 重载事件过滤器
 //
 // @route(reload_event_filter)
@@ -1824,22 +2130,28 @@ func (bot *CQBot) CQReloadEventFilter(file string) global.MSG {
 	return OK(nil)
 }
 
+// CQGetSupportedActions 获取支持的动作列表
+//
+// @route(get_supported_actions)
+func (bot *CQBot) CQGetSupportedActions(spec *onebot.Spec) global.MSG {
+	return OK(spec.SupportedActions)
+}
+
 // OK 生成成功返回值
-func OK(data interface{}) global.MSG {
-	return global.MSG{"data": data, "retcode": 0, "status": "ok"}
+func OK(data any) global.MSG {
+	return global.MSG{"data": data, "retcode": 0, "status": "ok", "message": ""}
 }
 
 // Failed 生成失败返回值
 func Failed(code int, msg ...string) global.MSG {
-	m := ""
-	w := ""
+	m, w := "", ""
 	if len(msg) > 0 {
 		m = msg[0]
 	}
 	if len(msg) > 1 {
 		w = msg[1]
 	}
-	return global.MSG{"data": nil, "retcode": code, "msg": m, "wording": w, "status": "failed"}
+	return global.MSG{"data": nil, "retcode": code, "msg": m, "wording": w, "message": w, "status": "failed"}
 }
 
 func limitedString(str string) string {
