@@ -108,7 +108,7 @@ func (hook *LocalHook) SetPath(path string) {
 }
 
 // NewLocalHook 初始化本地日志钩子实现
-func NewLocalHook(args interface{}, consoleFormatter, fileFormatter logrus.Formatter, levels ...logrus.Level) *LocalHook {
+func NewLocalHook(args any, consoleFormatter, fileFormatter logrus.Formatter, levels ...logrus.Level) *LocalHook {
 	hook := &LocalHook{
 		lock: new(sync.Mutex),
 	}
@@ -195,7 +195,8 @@ func (f LogFormat) Format(entry *logrus.Entry) ([]byte, error) {
 		buf.WriteString(colorReset)
 	}
 
-	ret := append([]byte(nil), buf.Bytes()...) // copy buffer
+	ret := make([]byte, len(buf.Bytes()))
+	copy(ret, buf.Bytes()) // copy buffer
 	return ret, nil
 }
 
