@@ -3,13 +3,10 @@ package param
 
 import (
 	"math"
-	"reflect"
 	"regexp"
 	"strings"
 	"sync"
 
-	"github.com/Mrs4s/MiraiGo/utils"
-	"github.com/segmentio/asm/base64"
 	"github.com/tidwall/gjson"
 )
 
@@ -22,7 +19,7 @@ import (
 // type gjson.True or gjson.False
 //
 // type string "true","yes","1" or "false","no","0" (case insensitive)
-func EnsureBool(p interface{}, defaultVal bool) bool {
+func EnsureBool(p any, defaultVal bool) bool {
 	var str string
 	if b, ok := p.(bool); ok {
 		return b
@@ -31,7 +28,7 @@ func EnsureBool(p interface{}, defaultVal bool) bool {
 		if !j.Exists() {
 			return defaultVal
 		}
-		switch j.Type { // nolint
+		switch j.Type { // nolint: exhaustive
 		case gjson.True:
 			return true
 		case gjson.False:
@@ -83,48 +80,4 @@ func SplitURL(s string) []string {
 	}
 	result = append(result, s[last:])
 	return result
-}
-
-// Base64DecodeString decode base64 with avx2
-// see https://github.com/segmentio/asm/issues/50
-// avoid incorrect unsafe usage in origin library
-func Base64DecodeString(s string) ([]byte, error) {
-	e := base64.StdEncoding
-	dst := make([]byte, e.DecodedLen(len(s)))
-	n, err := e.Decode(dst, utils.S2B(s))
-	return dst[:n], err
-}
-
-// SetAtDefault 在变量 variable 为默认值 defaultValue 的时候修改为 value
-func SetAtDefault(variable, value, defaultValue interface{}) {
-	v := reflect.ValueOf(variable)
-	v2 := reflect.ValueOf(value)
-	if v.Kind() != reflect.Ptr || v.IsNil() {
-		return
-	}
-	v = v.Elem()
-	if v.Interface() != defaultValue {
-		return
-	}
-	if v.Kind() != v2.Kind() {
-		return
-	}
-	v.Set(v2)
-}
-
-// SetExcludeDefault 在目标值 value 不为默认值 defaultValue 时修改 variable 为 value
-func SetExcludeDefault(variable, value, defaultValue interface{}) {
-	v := reflect.ValueOf(variable)
-	v2 := reflect.ValueOf(value)
-	if v.Kind() != reflect.Ptr || v.IsNil() {
-		return
-	}
-	v = v.Elem()
-	if reflect.Indirect(v2).Interface() != defaultValue {
-		return
-	}
-	if v.Kind() != v2.Kind() {
-		return
-	}
-	v.Set(v2)
 }
