@@ -3,17 +3,22 @@ package gocq
 import (
 	"bufio"
 	"bytes"
+	"fmt"
+	"image"
+	"image/png"
 	"os"
 	"strings"
 	"time"
 
-	qrcodeTerminal "github.com/Baozisoftware/qrcode-terminal-go"
 	"github.com/Mrs4s/MiraiGo/client"
-	"github.com/gocq/qrcode"
+	"github.com/Mrs4s/MiraiGo/utils"
+	"github.com/mattn/go-colorable"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
+	"gopkg.ilharper.com/x/isatty"
 
 	"github.com/Mrs4s/go-cqhttp/global"
+	"github.com/Mrs4s/go-cqhttp/internal/download"
 )
 
 var console = bufio.NewReader(os.Stdin)
@@ -24,7 +29,7 @@ func readLine() (str string) {
 	return
 }
 
-func readLineTimeout(t time.Duration, de string) (str string) {
+func readLineTimeout(t time.Duration) {
 	r := make(chan string)
 	go func() {
 		select {
@@ -32,15 +37,22 @@ func readLineTimeout(t time.Duration, de string) (str string) {
 		case <-time.After(t):
 		}
 	}()
-	str = de
 	select {
-	case str = <-r:
+	case <-r:
 	case <-time.After(t):
 	}
-	return
+}
+
+func readIfTTY(de string) (str string) {
+	if isatty.Isatty(os.Stdin.Fd()) {
+		return readLine()
+	}
+	log.Warnf("未检测到输入终端，自动选择%s.", de)
+	return de
 }
 
 var cli *client.QQClient
+var device *client.DeviceInfo
 
 // ErrSMSRequestError SMS请求出错
 var ErrSMSRequestError = errors.New("sms request error")
@@ -53,12 +65,36 @@ func commonLogin() error {
 	return loginResponseProcessor(res)
 }
 
-func qrcodeLogin() error {
-	rsp, err := cli.FetchQRCode()
+func printQRCode(imgData []byte) {
+	const (
+		black = "\033[48;5;0m  \033[0m"
+		white = "\033[48;5;7m  \033[0m"
+	)
+	img, err := png.Decode(bytes.NewReader(imgData))
 	if err != nil {
-		return err
+		log.Panic(err)
 	}
-	fi, err := qrcode.Decode(bytes.NewReader(rsp.ImageData))
+	data := img.(*image.Gray).Pix
+	bound := img.Bounds().Max.X
+	buf := make([]byte, 0, (bound*4+1)*(bound))
+	i := 0
+	for y := 0; y < bound; y++ {
+		i = y * bound
+		for x := 0; x < bound; x++ {
+			if data[i] != 255 {
+				buf = append(buf, white...)
+			} else {
+				buf = append(buf, black...)
+			}
+			i++
+		}
+		buf = append(buf, '\n')
+	}
+	_, _ = colorable.NewColorableStdout().Write(buf)
+}
+
+func qrcodeLogin() error {
+	rsp, err := cli.FetchQRCodeCustomSize(1, 2, 1)
 	if err != nil {
 		return err
 	}
@@ -70,7 +106,7 @@ func qrcodeLogin() error {
 		log.Infof("请使用手机QQ扫描二维码 (qrcode.png) : ")
 	}
 	time.Sleep(time.Second)
-	qrcodeTerminal.New2(qrcodeTerminal.ConsoleColors.BrightBlack, qrcodeTerminal.ConsoleColors.BrightWhite, qrcodeTerminal.QRCodeRecoveryLevels.Low).Get(fi.Content).Print()
+	printQRCode(rsp.ImageData)
 	s, err := cli.QueryQRCodeStatus(rsp.Sig)
 	if err != nil {
 		return err
@@ -117,11 +153,15 @@ func loginResponseProcessor(res *client.LoginResponse) error {
 		var text string
 		switch res.Error {
 		case client.SliderNeededError:
-			log.Warnf("登录需要滑条验证码, 请使用手机QQ扫描二维码以继续登录.")
-			cli.Disconnect()
-			cli.Release()
-			cli = client.NewClientEmpty()
-			return qrcodeLogin()
+			log.Warnf("登录需要滑条验证码, 请验证后重试.")
+			ticket := getTicket(res.VerifyUrl)
+			if ticket == "" {
+				log.Infof("按 Enter 继续....")
+				readLine()
+				os.Exit(0)
+			}
+			res, err = cli.SubmitTicket(ticket)
+			continue
 		case client.NeedCaptcha:
 			log.Warnf("登录需要验证码.")
 			_ = os.WriteFile("captcha.jpg", res.CaptchaImage, 0o644)
@@ -145,8 +185,8 @@ func loginResponseProcessor(res *client.LoginResponse) error {
 			log.Warnf("账号已开启设备锁，请选择验证方式:")
 			log.Warnf("1. 向手机 %v 发送短信验证码", res.SMSPhone)
 			log.Warnf("2. 使用手机QQ扫码验证.")
-			log.Warn("请输入(1 - 2) (将在10秒后自动选择2)：")
-			text = readLineTimeout(time.Second*10, "2")
+			log.Warn("请输入(1 - 2)：")
+			text = readIfTTY("2")
 			if strings.Contains(text, "1") {
 				if !cli.RequestSMS() {
 					log.Warnf("发送验证码失败，可能是请求过于频繁.")
@@ -161,20 +201,62 @@ func loginResponseProcessor(res *client.LoginResponse) error {
 		case client.UnsafeDeviceError:
 			log.Warnf("账号已开启设备锁，请前往 -> %v <- 验证后重启Bot.", res.VerifyUrl)
 			log.Infof("按 Enter 或等待 5s 后继续....")
-			readLineTimeout(time.Second*5, "")
+			readLineTimeout(time.Second * 5)
 			os.Exit(0)
 		case client.OtherLoginError, client.UnknownLoginError, client.TooManySMSRequestError:
 			msg := res.ErrorMessage
-			if strings.Contains(msg, "版本") {
-				msg = "密码错误或账号被冻结"
+			log.Warnf("登录失败: %v Code: %v", msg, res.Code)
+			switch res.Code {
+			case 235:
+				log.Warnf("设备信息被封禁, 请删除 device.json 后重试.")
+			case 237:
+				log.Warnf("登录过于频繁, 请在手机QQ登录并根据提示完成认证后等一段时间重试")
+			case 45:
+				log.Warnf("你的账号被限制登录, 请配置 SignServer 后重试")
 			}
-			if strings.Contains(msg, "冻结") {
-				log.Fatalf("账号被冻结")
-			}
-			log.Warnf("登录失败: %v", msg)
-			log.Infof("按 Enter 或等待 5s 后继续....")
-			readLineTimeout(time.Second*5, "")
+			log.Infof("按 Enter 继续....")
+			readLine()
 			os.Exit(0)
 		}
 	}
+}
+
+func getTicket(u string) string {
+	log.Warnf("请选择提交滑块ticket方式:")
+	log.Warnf("1. 自动提交")
+	log.Warnf("2. 手动抓取提交")
+	log.Warn("请输入(1 - 2)：")
+	text := readLine()
+	id := utils.RandomString(8)
+	auto := !strings.Contains(text, "2")
+	if auto {
+		u = strings.ReplaceAll(u, "https://ssl.captcha.qq.com/template/wireless_mqq_captcha.html?", fmt.Sprintf("https://captcha.go-cqhttp.org/captcha?id=%v&", id))
+	}
+	log.Warnf("请前往该地址验证 -> %v ", u)
+	if !auto {
+		log.Warn("请输入ticket： (Enter 提交)")
+		return readLine()
+	}
+
+	for count := 120; count > 0; count-- {
+		str := fetchCaptcha(id)
+		if str != "" {
+			return str
+		}
+		time.Sleep(time.Second)
+	}
+	log.Warnf("验证超时")
+	return ""
+}
+
+func fetchCaptcha(id string) string {
+	g, err := download.Request{URL: "https://captcha.go-cqhttp.org/captcha/ticket?id=" + id}.JSON()
+	if err != nil {
+		log.Debugf("获取 Ticket 时出现错误: %v", err)
+		return ""
+	}
+	if g.Get("ticket").Exists() {
+		return g.Get("ticket").String()
+	}
+	return ""
 }
