@@ -1,84 +1,75 @@
 package leveldb
 
 import (
-	"bytes"
-	"encoding/gob"
 	"path"
 
 	"github.com/Mrs4s/MiraiGo/binary"
+	"github.com/Mrs4s/MiraiGo/utils"
 	"github.com/pkg/errors"
 	"github.com/syndtr/goleveldb/leveldb"
 	"github.com/syndtr/goleveldb/leveldb/opt"
 	"gopkg.in/yaml.v3"
 
 	"github.com/Mrs4s/go-cqhttp/db"
-	"github.com/Mrs4s/go-cqhttp/global"
-	"github.com/Mrs4s/go-cqhttp/modules/config"
 )
 
-type LevelDBImpl struct {
+type database struct {
 	db *leveldb.DB
 }
 
-const (
-	group   byte = 0x0
-	private byte = 0x1
-)
+// config leveldb 相关配置
+type config struct {
+	Enable bool `yaml:"enable"`
+}
 
 func init() {
-	gob.Register(db.StoredMessageAttribute{})
-	gob.Register(db.QuotedInfo{})
-	gob.Register(global.MSG{})
-	gob.Register(db.StoredGroupMessage{})
-	gob.Register(db.StoredPrivateMessage{})
-
 	db.Register("leveldb", func(node yaml.Node) db.Database {
-		conf := new(config.LevelDBConfig)
+		conf := new(config)
 		_ = node.Decode(conf)
 		if !conf.Enable {
 			return nil
 		}
-		return &LevelDBImpl{}
+		return &database{}
 	})
 }
 
-func (ldb *LevelDBImpl) Open() error {
-	p := path.Join("data", "leveldb-v2")
+func (ldb *database) Open() error {
+	p := path.Join("data", "leveldb-v3")
 	d, err := leveldb.OpenFile(p, &opt.Options{
-		WriteBuffer: 128 * opt.KiB,
+		WriteBuffer: 32 * opt.KiB,
 	})
 	if err != nil {
-		return errors.Wrap(err, "open level ldb error")
+		return errors.Wrap(err, "open leveldb error")
 	}
 	ldb.db = d
 	return nil
 }
 
-func (ldb *LevelDBImpl) GetMessageByGlobalID(id int32) (db.StoredMessage, error) {
+func (ldb *database) GetMessageByGlobalID(id int32) (_ db.StoredMessage, err error) {
 	v, err := ldb.db.Get(binary.ToBytes(id), nil)
-	if err != nil {
+	if err != nil || len(v) == 0 {
 		return nil, errors.Wrap(err, "get value error")
 	}
-	r := binary.NewReader(v)
-	switch r.ReadByte() {
+	defer func() {
+		if r := recover(); r != nil {
+			err = errors.Errorf("%v", r)
+		}
+	}()
+	r, err := newReader(utils.B2S(v))
+	if err != nil {
+		return nil, err
+	}
+	switch r.uvarint() {
 	case group:
-		g := &db.StoredGroupMessage{}
-		if err = gob.NewDecoder(bytes.NewReader(r.ReadAvailable())).Decode(g); err != nil {
-			return nil, errors.Wrap(err, "decode message error")
-		}
-		return g, nil
+		return r.readStoredGroupMessage(), nil
 	case private:
-		p := &db.StoredPrivateMessage{}
-		if err = gob.NewDecoder(bytes.NewReader(r.ReadAvailable())).Decode(p); err != nil {
-			return nil, errors.Wrap(err, "decode message error")
-		}
-		return p, nil
+		return r.readStoredPrivateMessage(), nil
 	default:
 		return nil, errors.New("unknown message flag")
 	}
 }
 
-func (ldb *LevelDBImpl) GetGroupMessageByGlobalID(id int32) (*db.StoredGroupMessage, error) {
+func (ldb *database) GetGroupMessageByGlobalID(id int32) (*db.StoredGroupMessage, error) {
 	i, err := ldb.GetMessageByGlobalID(id)
 	if err != nil {
 		return nil, err
@@ -90,7 +81,7 @@ func (ldb *LevelDBImpl) GetGroupMessageByGlobalID(id int32) (*db.StoredGroupMess
 	return g, nil
 }
 
-func (ldb *LevelDBImpl) GetPrivateMessageByGlobalID(id int32) (*db.StoredPrivateMessage, error) {
+func (ldb *database) GetPrivateMessageByGlobalID(id int32) (*db.StoredPrivateMessage, error) {
 	i, err := ldb.GetMessageByGlobalID(id)
 	if err != nil {
 		return nil, err
@@ -102,28 +93,48 @@ func (ldb *LevelDBImpl) GetPrivateMessageByGlobalID(id int32) (*db.StoredPrivate
 	return p, nil
 }
 
-func (ldb *LevelDBImpl) InsertGroupMessage(msg *db.StoredGroupMessage) error {
-	buf := global.NewBuffer()
-	defer global.PutBuffer(buf)
-	if err := gob.NewEncoder(buf).Encode(msg); err != nil {
-		return errors.Wrap(err, "encode message error")
+func (ldb *database) GetGuildChannelMessageByID(id string) (*db.StoredGuildChannelMessage, error) {
+	v, err := ldb.db.Get([]byte(id), nil)
+	if err != nil {
+		return nil, errors.Wrap(err, "get value error")
 	}
-	err := ldb.db.Put(binary.ToBytes(msg.GlobalID), binary.NewWriterF(func(w *binary.Writer) {
-		w.WriteByte(group)
-		w.Write(buf.Bytes())
-	}), nil)
+	defer func() {
+		if r := recover(); r != nil {
+			err = errors.Errorf("%v", r)
+		}
+	}()
+	r, err := newReader(utils.B2S(v))
+	if err != nil {
+		return nil, err
+	}
+	switch r.uvarint() {
+	case guildChannel:
+		return r.readStoredGuildChannelMessage(), nil
+	default:
+		return nil, errors.New("unknown message flag")
+	}
+}
+
+func (ldb *database) InsertGroupMessage(msg *db.StoredGroupMessage) error {
+	w := newWriter()
+	w.uvarint(group)
+	w.writeStoredGroupMessage(msg)
+	err := ldb.db.Put(binary.ToBytes(msg.GlobalID), w.bytes(), nil)
 	return errors.Wrap(err, "put data error")
 }
 
-func (ldb *LevelDBImpl) InsertPrivateMessage(msg *db.StoredPrivateMessage) error {
-	buf := global.NewBuffer()
-	defer global.PutBuffer(buf)
-	if err := gob.NewEncoder(buf).Encode(msg); err != nil {
-		return errors.Wrap(err, "encode message error")
-	}
-	err := ldb.db.Put(binary.ToBytes(msg.GlobalID), binary.NewWriterF(func(w *binary.Writer) {
-		w.WriteByte(private)
-		w.Write(buf.Bytes())
-	}), nil)
+func (ldb *database) InsertPrivateMessage(msg *db.StoredPrivateMessage) error {
+	w := newWriter()
+	w.uvarint(private)
+	w.writeStoredPrivateMessage(msg)
+	err := ldb.db.Put(binary.ToBytes(msg.GlobalID), w.bytes(), nil)
+	return errors.Wrap(err, "put data error")
+}
+
+func (ldb *database) InsertGuildChannelMessage(msg *db.StoredGuildChannelMessage) error {
+	w := newWriter()
+	w.uvarint(guildChannel)
+	w.writeStoredGuildChannelMessage(msg)
+	err := ldb.db.Put(utils.S2B(msg.ID), w.bytes(), nil)
 	return errors.Wrap(err, "put data error")
 }
