@@ -10,23 +10,30 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/Mrs4s/go-cqhttp/db"
-	"github.com/Mrs4s/go-cqhttp/modules/config"
 )
 
-type MongoDBImpl struct {
+type database struct {
 	uri   string
 	db    string
 	mongo *mongo.Database
 }
 
+// config mongodb 相关配置
+type config struct {
+	Enable   bool   `yaml:"enable"`
+	URI      string `yaml:"uri"`
+	Database string `yaml:"database"`
+}
+
 const (
-	MongoGroupMessageCollection   = "group-messages"
-	MongoPrivateMessageCollection = "private-messages"
+	MongoGroupMessageCollection        = "group-messages"
+	MongoPrivateMessageCollection      = "private-messages"
+	MongoGuildChannelMessageCollection = "guild-channel-messages"
 )
 
 func init() {
-	db.Register("mongodb", func(node yaml.Node) db.Database {
-		conf := new(config.MongoDBConfig)
+	db.Register("database", func(node yaml.Node) db.Database {
+		conf := new(config)
 		_ = node.Decode(conf)
 		if conf.Database == "" {
 			conf.Database = "gocq-database"
@@ -34,11 +41,11 @@ func init() {
 		if !conf.Enable {
 			return nil
 		}
-		return &MongoDBImpl{uri: conf.URI, db: conf.Database}
+		return &database{uri: conf.URI, db: conf.Database}
 	})
 }
 
-func (m *MongoDBImpl) Open() error {
+func (m *database) Open() error {
 	cli, err := mongo.Connect(context.Background(), options.Client().ApplyURI(m.uri))
 	if err != nil {
 		return errors.Wrap(err, "open mongo connection error")
@@ -47,14 +54,14 @@ func (m *MongoDBImpl) Open() error {
 	return nil
 }
 
-func (m *MongoDBImpl) GetMessageByGlobalID(id int32) (db.StoredMessage, error) {
+func (m *database) GetMessageByGlobalID(id int32) (db.StoredMessage, error) {
 	if r, err := m.GetGroupMessageByGlobalID(id); err == nil {
 		return r, nil
 	}
 	return m.GetPrivateMessageByGlobalID(id)
 }
 
-func (m *MongoDBImpl) GetGroupMessageByGlobalID(id int32) (*db.StoredGroupMessage, error) {
+func (m *database) GetGroupMessageByGlobalID(id int32) (*db.StoredGroupMessage, error) {
 	coll := m.mongo.Collection(MongoGroupMessageCollection)
 	var ret db.StoredGroupMessage
 	if err := coll.FindOne(context.Background(), bson.D{{"globalId", id}}).Decode(&ret); err != nil {
@@ -63,7 +70,7 @@ func (m *MongoDBImpl) GetGroupMessageByGlobalID(id int32) (*db.StoredGroupMessag
 	return &ret, nil
 }
 
-func (m *MongoDBImpl) GetPrivateMessageByGlobalID(id int32) (*db.StoredPrivateMessage, error) {
+func (m *database) GetPrivateMessageByGlobalID(id int32) (*db.StoredPrivateMessage, error) {
 	coll := m.mongo.Collection(MongoPrivateMessageCollection)
 	var ret db.StoredPrivateMessage
 	if err := coll.FindOne(context.Background(), bson.D{{"globalId", id}}).Decode(&ret); err != nil {
@@ -72,14 +79,29 @@ func (m *MongoDBImpl) GetPrivateMessageByGlobalID(id int32) (*db.StoredPrivateMe
 	return &ret, nil
 }
 
-func (m *MongoDBImpl) InsertGroupMessage(msg *db.StoredGroupMessage) error {
+func (m *database) GetGuildChannelMessageByID(id string) (*db.StoredGuildChannelMessage, error) {
+	coll := m.mongo.Collection(MongoGuildChannelMessageCollection)
+	var ret db.StoredGuildChannelMessage
+	if err := coll.FindOne(context.Background(), bson.D{{"_id", id}}).Decode(&ret); err != nil {
+		return nil, errors.Wrap(err, "query error")
+	}
+	return &ret, nil
+}
+
+func (m *database) InsertGroupMessage(msg *db.StoredGroupMessage) error {
 	coll := m.mongo.Collection(MongoGroupMessageCollection)
-	_, err := coll.InsertOne(context.Background(), msg)
+	_, err := coll.UpdateOne(context.Background(), bson.D{{"_id", msg.ID}}, bson.D{{"$set", msg}}, options.Update().SetUpsert(true))
 	return errors.Wrap(err, "insert error")
 }
 
-func (m *MongoDBImpl) InsertPrivateMessage(msg *db.StoredPrivateMessage) error {
+func (m *database) InsertPrivateMessage(msg *db.StoredPrivateMessage) error {
 	coll := m.mongo.Collection(MongoPrivateMessageCollection)
-	_, err := coll.InsertOne(context.Background(), msg)
+	_, err := coll.UpdateOne(context.Background(), bson.D{{"_id", msg.ID}}, bson.D{{"$set", msg}}, options.Update().SetUpsert(true))
+	return errors.Wrap(err, "insert error")
+}
+
+func (m *database) InsertGuildChannelMessage(msg *db.StoredGuildChannelMessage) error {
+	coll := m.mongo.Collection(MongoGuildChannelMessageCollection)
+	_, err := coll.UpdateOne(context.Background(), bson.D{{"_id", msg.ID}}, bson.D{{"$set", msg}}, options.Update().SetUpsert(true))
 	return errors.Wrap(err, "insert error")
 }
