@@ -33,9 +33,15 @@ func EncoderSilk(data []byte) ([]byte, error) {
 // EncodeMP4 将给定视频文件编码为MP4
 func EncodeMP4(src string, dst string) error { //        -y 覆盖文件
 	cmd1 := exec.Command("ffmpeg", "-i", src, "-y", "-c", "copy", "-map", "0", dst)
+	if errors.Is(cmd1.Err, exec.ErrDot) {
+		cmd1.Err = nil
+	}
 	err := cmd1.Run()
 	if err != nil {
 		cmd2 := exec.Command("ffmpeg", "-i", src, "-y", "-c:v", "h264", "-c:a", "mp3", dst)
+		if errors.Is(cmd2.Err, exec.ErrDot) {
+			cmd2.Err = nil
+		}
 		return errors.Wrap(cmd2.Run(), "convert mp4 failed")
 	}
 	return err
@@ -43,6 +49,9 @@ func EncodeMP4(src string, dst string) error { //        -y 覆盖文件
 
 // ExtractCover 获取给定视频文件的Cover
 func ExtractCover(src string, target string) error {
-	cmd := exec.Command("ffmpeg", "-i", src, "-y", "-r", "1", "-f", "image2", target)
+	cmd := exec.Command("ffmpeg", "-i", src, "-y", "-ss", "0", "-frames:v", "1", target)
+	if errors.Is(cmd.Err, exec.ErrDot) {
+		cmd.Err = nil
+	}
 	return errors.Wrap(cmd.Run(), "extract video cover failed")
 }
