@@ -2,20 +2,10 @@
 package cache
 
 import (
-	"fmt"
-	"sync"
-
 	log "github.com/sirupsen/logrus"
-
-	"github.com/Mrs4s/go-cqhttp/global"
-	"github.com/Mrs4s/go-cqhttp/internal/base"
-	"github.com/Mrs4s/go-cqhttp/internal/btree"
+	"github.com/syndtr/goleveldb/leveldb"
+	"github.com/syndtr/goleveldb/leveldb/opt"
 )
-
-// todo(wdvxdr): always enable db-cache in v1.0.0
-
-// EnableCacheDB 是否启用 btree db缓存图片等
-var EnableCacheDB bool
 
 // Media Cache DBs
 var (
@@ -26,72 +16,36 @@ var (
 
 // Cache wraps the btree.DB for concurrent safe
 type Cache struct {
-	lock sync.RWMutex
-	db   *btree.DB
+	ldb *leveldb.DB
 }
 
 // Insert 添加媒体缓存
 func (c *Cache) Insert(md5, data []byte) {
-	c.lock.Lock()
-	defer c.lock.Unlock()
-
-	var hash [16]byte
-	copy(hash[:], md5)
-	c.db.Insert(&hash[0], data)
+	_ = c.ldb.Put(md5, data, nil)
 }
 
 // Get 获取缓存信息
 func (c *Cache) Get(md5 []byte) []byte {
-	c.lock.RLock()
-	defer c.lock.RUnlock()
-
-	var hash [16]byte
-	copy(hash[:], md5)
-	return c.db.Get(&hash[0])
+	got, _ := c.ldb.Get(md5, nil)
+	return got
 }
 
 // Delete 删除指定缓存
 func (c *Cache) Delete(md5 []byte) {
-	c.lock.Lock()
-	defer c.lock.Unlock()
-
-	var hash [16]byte
-	copy(hash[:], md5)
-	_ = c.db.Delete(&hash[0])
+	_ = c.ldb.Delete(md5, nil)
 }
 
 // Init 初始化 Cache
 func Init() {
-	node, ok := base.Database["cache"]
-	if !ok {
-		return
-	}
-	EnableCacheDB = true
-	var conf map[string]string
-	err := node.Decode(&conf)
-	if err != nil {
-		log.Fatalf("failed to read cache config: %v", err)
-	}
-
-	var open = func(typ string, cache *Cache) {
-		file := conf[typ]
-		if file == "" {
-			file = fmt.Sprintf("data/%s.db", typ)
+	open := func(typ, path string, cache *Cache) {
+		ldb, err := leveldb.OpenFile(path, &opt.Options{
+			WriteBuffer: 4 * opt.KiB,
+		})
+		if err != nil {
+			log.Fatalf("open cache %s db failed: %v", typ, err)
 		}
-		if global.PathExists(file) {
-			db, err := btree.Open(file)
-			if err != nil {
-				log.Fatalf("open %s cache failed: %v", typ, err)
-			}
-			cache.db = db
-		} else {
-			db, err := btree.Create(file)
-			if err != nil {
-				log.Fatalf("create %s cache failed: %v", typ, err)
-			}
-			cache.db = db
-		}
+		cache.ldb = ldb
 	}
-	open("image", &Image)
-	open("video", &Video)
+	open("image", "data/images", &Image)
+	open("video", "data/videos", &Video)
 }
