@@ -5,25 +5,95 @@ package api
 import (
 	"github.com/Mrs4s/go-cqhttp/coolq"
 	"github.com/Mrs4s/go-cqhttp/global"
+	"github.com/Mrs4s/go-cqhttp/pkg/onebot"
 )
 
-func (c *Caller) call(action string, p Getter) global.MSG {
+func (c *Caller) call(action string, spec *onebot.Spec, p Getter) global.MSG {
+	if spec.Version == 11 {
+		switch action {
+		case ".handle_quick_operation":
+			p0 := p.Get("context")
+			p1 := p.Get("operation")
+			return c.bot.CQHandleQuickOperation(p0, p1)
+		case "can_send_image":
+			return c.bot.CQCanSendImage()
+		case "can_send_record":
+			return c.bot.CQCanSendRecord()
+		case "get_login_info":
+			return c.bot.CQGetLoginInfo()
+		case "get_stranger_info":
+			p0 := p.Get("user_id").Int()
+			return c.bot.CQGetStrangerInfo(p0)
+		case "get_version_info":
+			return c.bot.CQGetVersionInfo()
+		case "send_forward_msg":
+			p0 := p.Get("group_id").Int()
+			p1 := p.Get("user_id").Int()
+			p2 := p.Get("messages")
+			p3 := p.Get("message_type").String()
+			return c.bot.CQSendForwardMessage(p0, p1, p2, p3)
+		case "send_group_forward_msg":
+			p0 := p.Get("group_id").Int()
+			p1 := p.Get("messages")
+			return c.bot.CQSendGroupForwardMessage(p0, p1)
+		case "send_group_msg":
+			p0 := p.Get("group_id").Int()
+			p1 := p.Get("message")
+			p2 := p.Get("auto_escape").Bool()
+			return c.bot.CQSendGroupMessage(p0, p1, p2)
+		case "send_msg":
+			p0 := p.Get("group_id").Int()
+			p1 := p.Get("user_id").Int()
+			p2 := p.Get("message")
+			p3 := p.Get("message_type").String()
+			p4 := p.Get("auto_escape").Bool()
+			return c.bot.CQSendMessage(p0, p1, p2, p3, p4)
+		case "send_private_forward_msg":
+			p0 := p.Get("user_id").Int()
+			p1 := p.Get("messages")
+			return c.bot.CQSendPrivateForwardMessage(p0, p1)
+		case "send_private_msg":
+			p0 := p.Get("user_id").Int()
+			p1 := p.Get("group_id").Int()
+			p2 := p.Get("message")
+			p3 := p.Get("auto_escape").Bool()
+			return c.bot.CQSendPrivateMessage(p0, p1, p2, p3)
+		}
+	}
+	if spec.Version == 12 {
+		switch action {
+		case "get_self_info":
+			return c.bot.CQGetLoginInfo()
+		case "get_user_info":
+			p0 := p.Get("user_id").Int()
+			return c.bot.CQGetStrangerInfo(p0)
+		case "get_version":
+			return c.bot.CQGetVersion()
+		case "send_message":
+			p0 := p.Get("group_id").String()
+			p1 := p.Get("user_id").String()
+			p2 := p.Get("detail_type").String()
+			p3 := p.Get("message")
+			return c.bot.CQSendMessageV12(p0, p1, p2, p3)
+		}
+	}
 	switch action {
-	default:
-		return coolq.Failed(404, "API_NOT_FOUND", "API不存在")
 	case ".get_word_slices":
 		p0 := p.Get("content").String()
 		return c.bot.CQGetWordSlices(p0)
-	case ".handle_quick_operation":
-		p0 := p.Get("context")
-		p1 := p.Get("operation")
-		return c.bot.CQHandleQuickOperation(p0, p1)
+	case ".ocr_image", "ocr_image":
+		p0 := p.Get("image").String()
+		return c.bot.CQOcrImage(p0)
+	case "_del_group_notice":
+		p0 := p.Get("group_id").Int()
+		p1 := p.Get("notice_id").String()
+		return c.bot.CQDelGroupMemo(p0, p1)
+	case "_get_group_notice":
+		p0 := p.Get("group_id").Int()
+		return c.bot.CQGetGroupMemo(p0)
 	case "_get_model_show":
 		p0 := p.Get("model").String()
 		return c.bot.CQGetModelShow(p0)
-	case "_get_vip_info":
-		p0 := p.Get("user_id").Int()
-		return c.bot.CQGetVipInfo(p0)
 	case "_send_group_notice":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("content").String()
@@ -33,10 +103,6 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		p0 := p.Get("model").String()
 		p1 := p.Get("model_show").String()
 		return c.bot.CQSetModelShow(p0, p1)
-	case "can_send_image":
-		return c.bot.CQCanSendImage()
-	case "can_send_record":
-		return c.bot.CQCanSendRecord()
 	case "check_url_safely":
 		p0 := p.Get("url").String()
 		return c.bot.CQCheckURLSafely(p0)
@@ -48,7 +114,7 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 	case "create_guild_role":
 		p0 := p.Get("guild_id").Uint()
 		p1 := p.Get("name").String()
-		p2 := uint32(p.Get("color").Int())
+		p2 := uint32(p.Get("color").Uint())
 		p3 := p.Get("independent").Bool()
 		p4 := p.Get("initial_users")
 		return c.bot.CQCreateGuildRole(p0, p1, p2, p3, p4)
@@ -61,7 +127,7 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 	case "delete_group_file":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("file_id").String()
-		p2 := int32(p.Get("bus_id").Int())
+		p2 := int32(p.Get("[busid,bus_id].0").Int())
 		return c.bot.CQGroupFileDeleteFile(p0, p1, p2)
 	case "delete_group_folder":
 		p0 := p.Get("group_id").Int()
@@ -89,7 +155,7 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		p0 := p.Get("[message_id,id].0").String()
 		return c.bot.CQGetForwardMessage(p0)
 	case "get_friend_list":
-		return c.bot.CQGetFriendList()
+		return c.bot.CQGetFriendList(spec)
 	case "get_group_at_all_remain":
 		p0 := p.Get("group_id").Int()
 		return c.bot.CQGetAtAllRemain(p0)
@@ -99,7 +165,7 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 	case "get_group_file_url":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("file_id").String()
-		p2 := int32(p.Get("bus_id").Int())
+		p2 := int32(p.Get("[busid,bus_id].0").Int())
 		return c.bot.CQGetGroupFileURL(p0, p1, p2)
 	case "get_group_files_by_folder":
 		p0 := p.Get("group_id").Int()
@@ -112,10 +178,10 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 	case "get_group_info":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("no_cache").Bool()
-		return c.bot.CQGetGroupInfo(p0, p1)
+		return c.bot.CQGetGroupInfo(p0, p1, spec)
 	case "get_group_list":
 		p0 := p.Get("no_cache").Bool()
-		return c.bot.CQGetGroupList(p0)
+		return c.bot.CQGetGroupList(p0, spec)
 	case "get_group_member_info":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("user_id").Int()
@@ -140,9 +206,21 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		return c.bot.CQGetGuildChannelList(p0, p1)
 	case "get_guild_list":
 		return c.bot.CQGetGuildList()
+	case "get_guild_member_list":
+		p0 := p.Get("guild_id").Uint()
+		p1 := p.Get("next_token").String()
+		return c.bot.CQGetGuildMembers(p0, p1)
+	case "get_guild_member_profile":
+		p0 := p.Get("guild_id").Uint()
+		p1 := p.Get("user_id").Uint()
+		return c.bot.CQGetGuildMemberProfile(p0, p1)
 	case "get_guild_meta_by_guest":
 		p0 := p.Get("guild_id").Uint()
 		return c.bot.CQGetGuildMetaByGuest(p0)
+	case "get_guild_msg":
+		p0 := p.Get("message_id").String()
+		p1 := p.Get("no_cache").Bool()
+		return c.bot.CQGetGuildMessage(p0, p1)
 	case "get_guild_roles":
 		p0 := p.Get("guild_id").Uint()
 		return c.bot.CQGetGuildRoles(p0)
@@ -151,8 +229,6 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 	case "get_image":
 		p0 := p.Get("file").String()
 		return c.bot.CQGetImage(p0)
-	case "get_login_info":
-		return c.bot.CQGetLoginInfo()
 	case "get_msg":
 		p0 := int32(p.Get("message_id").Int())
 		return c.bot.CQGetMessage(p0)
@@ -160,57 +236,32 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		p0 := p.Get("no_cache").Bool()
 		return c.bot.CQGetOnlineClients(p0)
 	case "get_status":
-		return c.bot.CQGetStatus()
-	case "get_stranger_info":
-		p0 := p.Get("user_id").Int()
-		return c.bot.CQGetStrangerInfo(p0)
+		return c.bot.CQGetStatus(spec)
+	case "get_supported_actions":
+		return c.bot.CQGetSupportedActions(spec)
 	case "get_topic_channel_feeds":
 		p0 := p.Get("guild_id").Uint()
 		p1 := p.Get("channel_id").Uint()
 		return c.bot.CQGetTopicChannelFeeds(p0, p1)
 	case "get_unidirectional_friend_list":
 		return c.bot.CQGetUnidirectionalFriendList()
-	case "get_version_info":
-		return c.bot.CQGetVersionInfo()
 	case "mark_msg_as_read":
 		p0 := int32(p.Get("message_id").Int())
 		return c.bot.CQMarkMessageAsRead(p0)
-	case "ocr_image", ".ocr_image":
-		p0 := p.Get("image").String()
-		return c.bot.CQOcrImage(p0)
 	case "qidian_get_account_info":
 		return c.bot.CQGetQiDianAccountInfo()
 	case "reload_event_filter":
 		p0 := p.Get("file").String()
 		return c.bot.CQReloadEventFilter(p0)
-	case "send_group_forward_msg":
+	case "send_group_sign":
 		p0 := p.Get("group_id").Int()
-		p1 := p.Get("messages")
-		return c.bot.CQSendGroupForwardMessage(p0, p1)
-	case "send_group_msg":
-		p0 := p.Get("group_id").Int()
-		p1 := p.Get("message")
-		p2 := p.Get("auto_escape").Bool()
-		return c.bot.CQSendGroupMessage(p0, p1, p2)
+		return c.bot.CQSendGroupSign(p0)
 	case "send_guild_channel_msg":
 		p0 := p.Get("guild_id").Uint()
 		p1 := p.Get("channel_id").Uint()
 		p2 := p.Get("message")
 		p3 := p.Get("auto_escape").Bool()
 		return c.bot.CQSendGuildChannelMessage(p0, p1, p2, p3)
-	case "send_msg":
-		p0 := p.Get("group_id").Int()
-		p1 := p.Get("user_id").Int()
-		p2 := p.Get("message")
-		p3 := p.Get("message_type").String()
-		p4 := p.Get("auto_escape").Bool()
-		return c.bot.CQSendMessage(p0, p1, p2, p3, p4)
-	case "send_private_msg":
-		p0 := p.Get("user_id").Int()
-		p1 := p.Get("group_id").Int()
-		p2 := p.Get("message")
-		p3 := p.Get("auto_escape").Bool()
-		return c.bot.CQSendPrivateMessage(p0, p1, p2, p3)
 	case "set_essence_msg":
 		p0 := int32(p.Get("message_id").Int())
 		return c.bot.CQSetEssenceMessage(p0)
@@ -238,6 +289,13 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 			p2 = pt.Bool()
 		}
 		return c.bot.CQSetGroupAdmin(p0, p1, p2)
+	case "set_group_anonymous":
+		p0 := p.Get("group_id").Int()
+		p1 := true
+		if pt := p.Get("enable"); pt.Exists() {
+			p1 = pt.Bool()
+		}
+		return c.bot.CQSetGroupAnonymous(p0, p1)
 	case "set_group_anonymous_ban":
 		p0 := p.Get("group_id").Int()
 		p1 := p.Get("[anonymous_flag,anonymous.flag].0").String()
@@ -248,7 +306,7 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		p1 := p.Get("user_id").Int()
 		p2 := uint32(1800)
 		if pt := p.Get("duration"); pt.Exists() {
-			p2 = uint32(pt.Int())
+			p2 = uint32(pt.Uint())
 		}
 		return c.bot.CQSetGroupBan(p0, p1, p2)
 	case "set_group_card":
@@ -292,11 +350,18 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		p2 := p.Get("role_id").Uint()
 		p3 := p.Get("users")
 		return c.bot.CQSetGuildMemberRole(p0, p1, p2, p3)
+	case "set_qq_profile":
+		p0 := p.Get("nickname")
+		p1 := p.Get("company")
+		p2 := p.Get("email")
+		p3 := p.Get("college")
+		p4 := p.Get("personal_note")
+		return c.bot.CQSetQQProfile(p0, p1, p2, p3, p4)
 	case "update_guild_role":
 		p0 := p.Get("guild_id").Uint()
 		p1 := p.Get("role_id").Uint()
 		p2 := p.Get("name").String()
-		p3 := uint32(p.Get("color").Int())
+		p3 := uint32(p.Get("color").Uint())
 		p4 := p.Get("indepedent").Bool()
 		return c.bot.CQModifyRoleInGuild(p0, p1, p2, p3, p4)
 	case "upload_group_file":
@@ -305,5 +370,11 @@ func (c *Caller) call(action string, p Getter) global.MSG {
 		p2 := p.Get("name").String()
 		p3 := p.Get("folder").String()
 		return c.bot.CQUploadGroupFile(p0, p1, p2, p3)
+	case "upload_private_file":
+		p0 := p.Get("user_id").Int()
+		p1 := p.Get("file").String()
+		p2 := p.Get("name").String()
+		return c.bot.CQUploadPrivateFile(p0, p1, p2)
 	}
+	return coolq.Failed(404, "API_NOT_FOUND", "API不存在")
 }
